refactor(setupd): split main.rs into focused modules
Extract setup, update, UI, callback, and D-Bus logic into dedicated modules so the daemon is easier to navigate and amend without changing runtime behavior or test coverage.
