+++ components/feral-setupd/src/main.rs
@@ -1,375 +1,35 @@
+mod app_state;
 mod ble;
+mod callbacks;
 mod cdp;
 mod cfg;
 mod connectivity;
 mod constant;
+mod dbus_handlers;
 mod dbus_utils;
 mod encoding;
 mod log_uploader;
 mod persistent_state;
+mod phase_logic;
 mod setup_lifecycle;
+mod startup;
 mod system;
+mod ui;
+mod update_coordinator;
 mod updater;
 mod wifi_utils;
 
-use crate::persistent_state::PersistentState;
-use crate::setup_lifecycle::{SetupLifecycle, SetupPhase};
-use crate::wifi_utils::{Error as WifiError, SSIDsCacher};
-use anyhow::Context;
 use anyhow::Result;
-use ble::{Ble, BleCallbacks};
-use cdp::Cdp;
-use connectivity::Connectivity;
-use std::sync::Arc;
-use std::sync::atomic::{AtomicBool, Ordering};
-use std::time::Instant;
-use std::time::{SystemTime, UNIX_EPOCH};
-use tokio::signal::unix::{SignalKind, signal as unix_signal};
-use tokio::{
-    sync::{Mutex, mpsc},
-    task,
-    time::{self, Duration},
+use ble::Ble;
+use dbus_handlers::{setup_dbus_listeners, wait_for_controld};
+use startup::{
+    init_app_state, init_cdp, start_ble, startup_with_internet, startup_without_internet,
 };
-
-/// Controls when an update should be triggered.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum UpdateMode {
-    /// Update only when below the minimum supported version (mandatory update).
-    Required,
-    /// Update when any newer version is available (optional/user-triggered update).
-    Available,
-}
-
-/// Controls how the update check executes side effects (UI updates, update process).
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum UpdateExecution {
-    /// Run operations in foreground (blocking). Use for startup/D-Bus flows where
-    /// we can wait for completion.
-    Blocking,
-    /// Spawn operations in background. Use for BLE flows where we need to return
-    /// quickly to send a response to the mobile app.
-    NonBlocking,
-}
-
-#[inline]
-fn unix_s() -> i64 {
-    match SystemTime::now().duration_since(UNIX_EPOCH) {
-        Ok(duration) => duration.as_secs() as i64,
-        Err(error) => {
-            eprintln!("MAIN: System time is before UNIX_EPOCH: {error:?}");
-            0
-        }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-enum Page {
-    None(i64),
-    QRCode(i64),
-    Message(i64, String),
-    SystemUpgrade(i64),
-    FactoryReset(i64),
-    WebApp(i64),
-    ReflashingRequired(i64, String),
-}
-
-impl Page {
-    /// Check if the page should be kept when bluetooth disconnects
-    fn should_keep_on_bt_disconnect(&self) -> bool {
-        matches!(
-            self,
-            Page::WebApp(_)
-                | Page::SystemUpgrade(_)
-                | Page::FactoryReset(_)
-                | Page::ReflashingRequired(_, _)
-        )
-    }
-}
-
-#[derive(Debug)]
-struct AppState {
-    device_id: String,
-    branch: String,
-    current_version: String,
-    state_store: PersistentState,
-    internet: Connectivity,
-    page: Mutex<Page>,
-
-    // This is the flag to indicate whether we should automatically redirect to webapp
-    // when internet is available.
-    // On a second boot, if the internet is unavailable, users have 2 choices
-    // 1. Fix the internet connection, it will automatically check for update & play artwork
-    // 2. Scan the QRCode and start everything over again
-    // We need this flag to turn off the first flow if the user has chosen to provide a different wifi
-    // true = auto proceed, false = user has chosen to provide a different wifi
-    auto_proceed: AtomicBool,
-    /// Setup progress coordinator. Derives mobile-facing phase from persistent state
-    /// (PAIRED, topic_id), active operations (wifi, version check, update), and
-    /// permanent failure latch.
-    lifecycle: SetupLifecycle,
-    /// Single-flight guard for OTA updates to prevent concurrent update attempts.
-    /// Wrapped in Arc to allow UpdateGuard to be 'static for tokio::spawn.
-    update_in_progress: Arc<AtomicBool>,
-}
-
-const MAX_UPDATE_RETRIES: u8 = 3;
-
-// Message shown when update fails (fresh failure during current session)
-const UPDATE_FAILED_FRESH_MSG: &str = "Update didn't complete. Use your phone to try again, or restart the device. If it still fails, contact support@feralfile.com.";
-
-// Message shown on reboot when UpdateFailed phase is restored from disk.
-// User already restarted, so guide them to retry via phone or contact support.
-const UPDATE_FAILED_RECOVERED_MSG: &str = "Previous update failed. The device attempted to update but couldn't complete. Use your phone to retry the update. If the problem persists, contact support@feralfile.com for assistance.";
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum UpdateErrorType {
-    Transient,
-    Permanent,
-}
-
-/// Update failure with classification decided from the updater message before
-/// setupd adds context wrappers for logging.
-#[derive(Debug)]
-struct UpdateAttemptError {
-    source: anyhow::Error,
-    kind: UpdateErrorType,
-}
-
-impl std::fmt::Display for UpdateAttemptError {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        self.source.fmt(f)
-    }
-}
-
-impl std::error::Error for UpdateAttemptError {
-    fn source(&self) -> Option<&(dyn std::error::Error + 'static)> {
-        Some(self.source.as_ref())
-    }
-}
-
-/// Classify error messages from bash update scripts and updater infrastructure.
-/// Transient errors are worth retrying at the setupd level; permanent errors are not.
-///
-/// This classification matches actual error messages produced by:
-/// - feral-updater.sh and feral-system-update.sh (ffos repo)
-/// - updater.rs internal errors (this repo)
-///
-/// # ⚠️ FRAGILE CONTRACT WARNING
-///
-/// This function relies on **exact string matching** with error messages from shell scripts
-/// in the `ffos` repository (`feral-system-update.sh`, `feral-recovery-update.sh`).
-///
-/// **Any changes to script error messages MUST be accompanied by:**
-/// - Updates to this classification function
-/// - Verification via integration testing
-///
-/// **Future improvement:** Consider structured error codes or stable prefixes to reduce
-/// brittleness and improve cross-repository maintainability.
-fn classify_updater_message(message: &str) -> UpdateErrorType {
-    let msg = message.trim();
-
-    // ════════════════════════════════════════════════════════════════════════
-    // PERMANENT ERRORS: Do not retry at setupd level
-    // ════════════════════════════════════════════════════════════════════════
-
-    // Cryptographic failures (server published bad/corrupt image)
-    if msg.starts_with("Error: Signature verification failed for")
-        || msg.starts_with("Error: Signature file ")
-    {
-        return UpdateErrorType::Permanent;
-    }
-
-    // Image integrity failures (corrupt or invalid image structure)
-    if msg == "airootfs.sfs not found in image." || msg == "No post-extraction script found in ISO"
-    {
-        return UpdateErrorType::Permanent;
-    }
-
-    // Filesystem operation failures (disk full, permissions, etc.)
-    if msg == "Failed to create snapshot '/.snapshots/@ota_prev'. Aborting."
-        || msg.starts_with("Failed to create snapshot")
-    {
-        return UpdateErrorType::Permanent;
-    }
-
-    // Unknown/unclassified errors from updater.rs
-    if msg == "Unknown error occurred" {
-        return UpdateErrorType::Permanent;
-    }
-
-    // ════════════════════════════════════════════════════════════════════════
-    // TRANSIENT ERRORS: Worth retrying at setupd level
-    // ════════════════════════════════════════════════════════════════════════
-
-    // Network connectivity and download failures (no internet or server unreachable)
-    if msg == "No network connection. Aborting update."
-        || msg == "Failed to retrieve content length for image."
-        || msg == "Failed to download OTA image."
-        || msg == "Failed to download signature file."
-        || msg == "Failed to download recovery ISO."
-    {
-        return UpdateErrorType::Transient;
-    }
-
-    // setupd infrastructure failures (service start, log file access)
-    if msg.contains("Failed to start updater service")
-        || msg.contains("Failed to open /var/log/updaterd.log")
-        || msg == "updater closed channel without sending progress"
-    {
-        return UpdateErrorType::Transient;
-    }
-
-    // Lock failures (another update process running or stale lock)
-    // Transient because lock may be released by the time we retry
-    if msg.contains("Lock already held by another instance")
-        || msg.contains("either Lock already held")
-    {
-        return UpdateErrorType::Transient;
-    }
-
-    // ════════════════════════════════════════════════════════════════════════
-    // BASH ERR TRAP: Parse the trapped command to determine if network-related
-    // ════════════════════════════════════════════════════════════════════════
-    //
-    // Format: EXCEPTION ERR: LINE=118 CMD="curl -u ... -o /var/tmp/ota/image.zip"
-    //
-    // Network commands (curl, wget) are transient; other failures (mount, rsync,
-    // btrfs, mkinitcpio, etc.) are permanent.
-    //
-    if msg.starts_with("EXCEPTION ERR:") {
-        if msg.contains("curl") || msg.contains("wget") {
-            return UpdateErrorType::Transient;
-        }
-        return UpdateErrorType::Permanent;
-    }
-
-    // ════════════════════════════════════════════════════════════════════════
-    // DEFAULT: Unknown errors are permanent (avoid infinite retry loops)
-    // ════════════════════════════════════════════════════════════════════════
-    UpdateErrorType::Permanent
-}
-
-fn get_setup_phase(app_state: &AppState) -> String {
-    app_state.lifecycle.get().as_str().to_string()
-}
-
-/// Phase the BLE setup wrapper should leave behind when a version check ends without an update
-/// (failure or critical error). `check_and_update_system` already restored the pre-check durable
-/// phase (Ready/Pairing) before returning, so the wrapper must NOT blindly force Idle or it would
-/// demote a device that just recovered from `update_failed` via a BLE retry. Only non-durable
-/// phases collapse to Idle.
-fn ble_terminal_reset_phase(current: SetupPhase) -> SetupPhase {
-    if current.is_durable() {
-        current
-    } else {
-        SetupPhase::Idle
-    }
-}
-
-/// Whether the BLE success path must fetch a fresh relayer topic from controld.
-///
-/// `GetRelayerTopicID` blocks up to `DBUS_RELAYER_CHECK_TIMEOUT` (31s) and the BLE handler awaits
-/// the setup callback before replying to the phone, so an unconditional fetch defeats the bounded
-/// BLE response and can fail an already-provisioned device with `ServerUnreachable` even though it
-/// already holds the topic it should return. Only contact controld when no usable topic is
-/// persisted yet (first-time setup); a Pairing/Ready device returns its persisted topic immediately.
-fn needs_relayer_topic_fetch(persisted_topic: Option<&str>) -> bool {
-    !matches!(persisted_topic, Some(topic) if !topic.is_empty())
-}
-
-/// Whether a `show_pairing_qr_code` switch (QR or webapp) must be ignored.
-///
-/// A permanent update failure latches the durable `UpdateFailed` phase and leaves the failure
-/// message on screen until an explicit BLE/D-Bus retry clears it. `handle_permanent_update_failure`
-/// releases the update guard, so gating only on `update_in_progress` is not enough: a later QR/webapp
-/// switch from controld could navigate Chromium off the failure surface while `setup_phase` is still
-/// `update_failed`. Block both switch directions until recovery clears the latch.
-fn qr_switch_blocked_by_update_failed(phase: SetupPhase) -> bool {
-    phase == SetupPhase::UpdateFailed
-}
-
-/// Whether a `show_pairing_qr_code(false)` pairing confirmation should record the durable
-/// `Pairing -> Ready` transition for the current live phase.
-///
-/// The confirmation is a one-shot D-Bus signal (controld ACKs it after the callback returns and
-/// does not re-emit), so it must be recorded even if an OTA owns the device — otherwise gating the
-/// whole switch on the update guard silently drops the pairing confirmation and the device is
-/// stuck in `Pairing` while mobile believes pairing succeeded.
-///
-/// It applies when:
-/// - the device is in steady-state `Pairing` (the common case), or
-/// - an in-flight OTA op (guard held) has transiently masked `Pairing` with a non-durable phase
-///   while a relayer `topic_id` is already persisted. A persisted topic means the device is past
-///   topic allocation (it was `Pairing`/`Ready`, not a fresh `Idle`), so promoting to `Ready` is
-///   correct. Paired with [`phase_after_inflight_check`], which refuses to demote this `Ready`
-///   back to `Pairing` when the masking check restores.
-///
-/// It must NOT fire for `Idle` without a topic (first run), `Ready` (already paired), or
-/// `UpdateFailed` (latched failure surface, awaits explicit retry).
-fn pairing_confirmation_promotes_to_ready(phase: SetupPhase, has_topic: bool) -> bool {
-    match phase {
-        SetupPhase::Pairing => true,
-        SetupPhase::WifiConnecting | SetupPhase::CheckingVersion | SetupPhase::Updating => {
-            has_topic
-        }
-        SetupPhase::Idle | SetupPhase::Ready | SetupPhase::UpdateFailed => false,
-    }
-}
-
-/// Phase to restore after an in-flight no-update / failed version check, given the phase captured
-/// at the check's entry (`phase_before_check`) and the live phase now (`current_phase`).
-///
-/// Normally this restores the durable phase captured before the transient `CheckingVersion` (or
-/// `Idle` if that entry phase was not durable), preserving paired state across a no-op check.
-///
-/// The exception preserves a pairing confirmation that landed *during* the check: if the live phase
-/// was promoted to `Ready` while we were checking (see [`pairing_confirmation_promotes_to_ready`]),
-/// demoting it back to `Pairing` would silently drop the confirmation. So a `Ready` that appeared
-/// mid-check wins over a captured `Pairing`.
-fn phase_after_inflight_check(
-    phase_before_check: SetupPhase,
-    current_phase: SetupPhase,
-) -> SetupPhase {
-    let restore = if phase_before_check.is_durable() {
-        phase_before_check
-    } else {
-        SetupPhase::Idle
-    };
-
-    if restore == SetupPhase::Pairing && current_phase == SetupPhase::Ready {
-        SetupPhase::Ready
-    } else {
-        restore
-    }
-}
-
-/// RAII guard for the update_in_progress flag.
-/// Automatically releases the flag when dropped, ensuring cleanup on all return paths
-/// (success, early return, panic, or error propagation via `?`).
-///
-/// Owns an Arc<AtomicBool> to be 'static-safe for tokio::spawn usage.
-struct UpdateGuard {
-    flag: Arc<AtomicBool>,
-}
-
-impl UpdateGuard {
-    fn try_acquire(flag: &Arc<AtomicBool>) -> Option<Self> {
-        if flag
-            .compare_exchange(false, true, Ordering::Acquire, Ordering::Relaxed)
-            .is_ok()
-        {
-            Some(UpdateGuard { flag: flag.clone() })
-        } else {
-            None
-        }
-    }
-}
-
-impl Drop for UpdateGuard {
-    fn drop(&mut self) {
-        self.flag.store(false, Ordering::Release);
-    }
-}
+use std::sync::Arc;
+use std::sync::atomic::Ordering;
+use tokio::signal::unix::{SignalKind, signal as unix_signal};
+use tokio::time::Duration;
+use wifi_utils::SSIDsCacher;
 
 // Sentry has to start before tokio runtime
 // That's why we can't use #[tokio::main]
@@ -451,7 +111,7 @@ async fn run() -> Result<()> {
     }
 
     // Wait for Ctrl+C or shutdown event
-    wait_for_shutdown().await; // Ignore any errors
+    wait_for_shutdown().await;
     println!("MAIN: Shutting down...");
     println!("MAIN: Stopping DBus listener...");
     stop_dbus_listener.store(true, Ordering::Relaxed);
@@ -464,1074 +124,6 @@ async fn run() -> Result<()> {
     }
     println!("MAIN: Shutting down...");
     Ok(())
-}
-
-async fn init_app_state(ble_service: &Arc<Ble>) -> Result<Arc<AppState>> {
-    let lifecycle = SetupLifecycle::new();
-    let state_store = PersistentState::new(constant::CACHE_FILEPATH)?;
-
-    // Restore durable phase (Pairing, Ready, UpdateFailed) from persistent storage after restart
-    lifecycle.restore_from_store(&state_store);
-
-    let app_state = Arc::new(AppState {
-        device_id: ble_service.get_device_id().await,
-        branch: cfg::branch().await?.to_string(),
-        current_version: cfg::current_version().await?.to_string(),
-        state_store,
-        internet: Connectivity::spawn().await,
-        page: Mutex::new(Page::None(unix_s())),
-        auto_proceed: AtomicBool::new(false),
-        lifecycle,
-        update_in_progress: Arc::new(AtomicBool::new(false)),
-    });
-    sentry::configure_scope(|scope| {
-        scope.set_tag("branch", app_state.branch.clone());
-        scope.set_tag("version", app_state.current_version.clone());
-        scope.set_user(Some(sentry::User {
-            id: Some(app_state.device_id.clone()),
-            ..Default::default()
-        }));
-    });
-    println!("MAIN: App state initialized: {app_state:?}");
-    Ok(app_state)
-}
-
-async fn init_cdp() -> Result<Arc<Cdp>> {
-    let chrome = Cdp::connect(constant::CDP_URL)
-        .await
-        .context("connecting to CDP")?;
-    Ok(Arc::new(chrome))
-}
-
-async fn start_ble(
-    ble_service: &Arc<Ble>,
-    app_state: &Arc<AppState>,
-    chrome: &Arc<Cdp>,
-    ssids_cacher: &Arc<SSIDsCacher>,
-) -> Result<()> {
-    let ble_callbacks = BleCallbacks {
-        bt_connected: callbacks::create_bt_connected_cb(app_state.clone(), chrome.clone()),
-        bt_disconnected: callbacks::create_bt_disconnected_cb(app_state.clone(), chrome.clone()),
-        factory_reset: callbacks::create_factory_reset_cb(app_state.clone(), chrome.clone()),
-        submit_logs: callbacks::create_submit_logs_cb(app_state.clone()),
-        connect_wifi: callbacks::create_connect_wifi_cb(app_state.clone(), chrome.clone()),
-        keep_wifi: callbacks::create_keep_wifi_cb(app_state.clone(), chrome.clone()),
-        get_info: callbacks::create_get_info_cb(app_state.clone()),
-    };
-
-    ble_service
-        .start(ble_callbacks, ssids_cacher.clone())
-        .await
-        .context("starting Bluetooth advertising")?;
-    println!("MAIN: Bluetooth advertising started successfully");
-    Ok(())
-}
-
-/// Whether a startup path must short-circuit to the `UpdateFailed` recovery screen instead of the
-/// normal QR / update-check flow.
-///
-/// Both the online (`on_startup_with_internet`) and offline (`startup_without_internet`) paths
-/// consult this so they stay in lockstep: an offline reboot must show the same failure UI as an
-/// online reboot. Only the durable `UpdateFailed` phase qualifies — any other phase proceeds with
-/// the normal flow. Keeping the decision in one predicate prevents the two branches from drifting.
-fn startup_requires_update_failed_recovery(phase: SetupPhase) -> bool {
-    phase == SetupPhase::UpdateFailed
-}
-
-/// Reboot-recovery UI for a persisted `UpdateFailed` phase.
-///
-/// Shows the recovered-failure message and leaves the phase untouched. Recovery is driven by an
-/// explicit BLE/D-Bus retry (which also re-establishes connectivity), so callers must `return`
-/// after this instead of falling through to QR / auto-proceed / update-check. Centralizing the
-/// surface keeps the online and offline startup branches identical if the message changes.
-async fn show_update_failed_recovery(app_state: &Arc<AppState>, chrome: &Arc<Cdp>) -> Result<()> {
-    println!("MAIN: UpdateFailed phase restored; showing recovered failure message");
-    println!("MAIN: Waiting for explicit BLE/D-Bus retry");
-    show_system_upgrade(chrome, app_state, UPDATE_FAILED_RECOVERED_MSG).await
-}
-
-/// Startup path when the device does **not** have internet at boot time.
-///
-/// When this is called:
-/// - `run` has already waited for `controld` to be reachable.
-/// - The initial internet check says the device is currently offline.
-///
-/// What it does:
-/// - Warms the Wi-Fi SSID cache so the first BLE scan is fast.
-/// - Shows the pairing QR code to let the user fix connectivity.
-/// - Polls for internet with an aggressive or relaxed interval depending on
-///   whether the device has ever connected before.
-/// - Marks the device as "has connected before" in the cache once online.
-/// - If the BLE flow has not opted out via `auto_proceed`, hands off to the
-///   normal "startup with internet" flow.
-///
-/// Notes:
-/// - If the user chooses a new Wi-Fi via BLE, the BLE flow clears
-///   `auto_proceed`; in that case this function will not auto-advance and the
-///   BLE setup path remains in control.
-async fn startup_without_internet(
-    app_state: &Arc<AppState>,
-    chrome: &Arc<Cdp>,
-    ssids_cacher: &Arc<SSIDsCacher>,
-    used_to_connect: Option<&String>,
-) -> Result<()> {
-    // Show the QRCode so the user can do something with the internet
-    let start_time = Instant::now();
-    let _ = ssids_cacher.get().await;
-    println!(
-        "MAIN: Get SSIDs in {:?} ms",
-        start_time.elapsed().as_millis()
-    );
-
-    // UpdateFailed reboot recovery must show the failure screen even when offline. Otherwise the
-    // device would display QR (and a BLE connect could swap in the welcome message) while still
-    // reporting setup_phase=update_failed to the mobile app. Recovery happens via an explicit
-    // BLE/D-Bus retry that brings its own connectivity, so we skip QR/poll/auto-proceed here. The
-    // SSID cache was warmed above so the recovery BLE scan stays fast.
-    if startup_requires_update_failed_recovery(app_state.lifecycle.get()) {
-        let _ = show_update_failed_recovery(app_state, chrome).await;
-        return Ok(());
-    }
-
-    let _ = show_qrcode(app_state, chrome).await;
-    app_state.auto_proceed.store(true, Ordering::Release);
-    // If somehow, the device has internet
-    // 1. Users fix the previous internet
-    // 2. Users plug in the LAN cable (instead of setting up wifi via bluetooth)
-    // We will take action immediately
-    let urgency = if used_to_connect.is_some() {
-        Duration::from_millis(constant::AGGRESSIVE_INTERNET_CHECK_INTERVAL)
-    } else {
-        Duration::from_millis(constant::RELAXED_INTERNET_CHECK_INTERVAL)
-    };
-    app_state.internet.wait_until_online(urgency, None).await;
-
-    if used_to_connect.is_none() {
-        app_state
-            .state_store
-            .set(persistent_state::CONNECTED, "true");
-        app_state.state_store.save()?;
-    }
-    // We now have internet, but we need to check if
-    // the internet comes from bluetooth (auto_proceed is set to false)
-    // if it's from bluetooth, we shouldn't do anything else as the bluetooth
-    // flow will handle it.
-    //
-    // Use compare_exchange so the auto-advance fires at most once and cannot race
-    // a concurrent BLE connect_wifi that clears auto_proceed right after we observe
-    // internet. Claiming the flag (true -> false) here also prevents a later BLE
-    // flow from re-triggering startup once we've taken ownership.
-    if app_state
-        .auto_proceed
-        .compare_exchange(true, false, Ordering::AcqRel, Ordering::Acquire)
-        .is_ok()
-    {
-        on_startup_with_internet(app_state.clone(), chrome.clone()).await?;
-    }
-    Ok(())
-}
-
-/// Startup path when the device already has internet at boot time.
-///
-/// When this is called:
-/// - `run` has already waited for `controld` to be reachable.
-/// - The initial internet check says the device is currently online.
-///
-/// What it does:
-/// - Ensures the "has ever connected" flag is persisted in the cache.
-/// - Delegates to `on_startup_with_internet` to either show the web app or a
-///   reflashing QR code, depending on updater state and cached topic ID.
-///
-/// Notes:
-/// - This path is used both on true first-boot with working internet and on
-///   subsequent boots where connectivity is available immediately.
-async fn startup_with_internet(
-    app_state: &Arc<AppState>,
-    chrome: &Arc<Cdp>,
-    used_to_connect: Option<&String>,
-) -> Result<()> {
-    if used_to_connect.is_none() {
-        app_state
-            .state_store
-            .set(persistent_state::CONNECTED, "true");
-        app_state.state_store.save()?;
-    }
-    on_startup_with_internet(app_state.clone(), chrome.clone()).await
-}
-
-async fn setup_dbus_listeners(
-    app_state: &Arc<AppState>,
-    chrome: &Arc<Cdp>,
-) -> Result<Arc<AtomicBool>> {
-    // Listen for QRCode switch signal
-    let qrcode_switch_cb = callbacks::create_qrcode_switch_cb(app_state.clone(), chrome.clone());
-    let stop_dbus_listener = Arc::new(AtomicBool::new(false));
-    dbus_utils::listen_for_signal(
-        constant::DBUS_CONTROLD_OBJECT,
-        constant::DBUS_CONTROLD_INTERFACE,
-        constant::DBUS_EVENT_QRCODE_SWITCH,
-        stop_dbus_listener.clone(),
-        qrcode_switch_cb,
-    )
-    .await?;
-
-    // Listen for factory reset signal
-    let factory_reset_cb =
-        callbacks::create_factory_reset_dbus_cb(app_state.clone(), chrome.clone());
-    dbus_utils::listen_for_signal(
-        constant::DBUS_CONTROLD_OBJECT,
-        constant::DBUS_CONTROLD_INTERFACE,
-        constant::DBUS_EVENT_FACTORY_RESET,
-        stop_dbus_listener.clone(),
-        factory_reset_cb,
-    )
-    .await?;
-
-    // Listen for upload logs signal
-    let upload_logs_cb = callbacks::create_upload_logs_dbus_cb(app_state.clone());
-    dbus_utils::listen_for_signal(
-        constant::DBUS_CONTROLD_OBJECT,
-        constant::DBUS_CONTROLD_INTERFACE,
-        constant::DBUS_EVENT_UPLOAD_LOGS,
-        stop_dbus_listener.clone(),
-        upload_logs_cb,
-    )
-    .await?;
-
-    // Listen for support-bundled upload logs signal
-    let upload_logs_with_bundle_cb =
-        callbacks::create_upload_logs_with_bundle_dbus_cb(app_state.clone());
-    dbus_utils::listen_for_signal(
-        constant::DBUS_CONTROLD_OBJECT,
-        constant::DBUS_CONTROLD_INTERFACE,
-        constant::DBUS_EVENT_UPLOAD_LOGS_WITH_BUNDLE,
-        stop_dbus_listener.clone(),
-        upload_logs_with_bundle_cb,
-    )
-    .await?;
-
-    // Listen for system update signal
-    let system_update_cb =
-        callbacks::create_system_update_dbus_cb(app_state.clone(), chrome.clone());
-    dbus_utils::listen_for_signal(
-        constant::DBUS_CONTROLD_OBJECT,
-        constant::DBUS_CONTROLD_INTERFACE,
-        constant::DBUS_EVENT_SYSTEM_UPDATE,
-        stop_dbus_listener.clone(),
-        system_update_cb,
-    )
-    .await?;
-
-    Ok(stop_dbus_listener)
-}
-
-async fn internet_setup_successfully_cb(
-    app_state: &Arc<AppState>,
-    chromium: &Arc<Cdp>,
-    guard: UpdateGuard,
-) -> Result<String, ble::BleStatus> {
-    // Check and update system using Required mode (only mandatory updates)
-    // Use NonBlocking execution since BLE flow needs to return quickly.
-    // `guard` was acquired by the BLE callback before any Wi-Fi/phase/UI side effects and is handed
-    // off here so the whole retry is serialized against OTA.
-    match check_and_update_system(
-        app_state,
-        chromium,
-        UpdateMode::Required,
-        UpdateExecution::NonBlocking,
-        guard,
-    )
-    .await
-    {
-        Ok(UpdateCheckResult::TooOldToUpgrade) => {
-            app_state.lifecycle.set(SetupPhase::Idle);
-            return Err(ble::BleStatus::VersionTooOld);
-        }
-        Ok(UpdateCheckResult::UpdateStarted) | Ok(UpdateCheckResult::UpdateInProgress) => {
-            return Err(ble::BleStatus::DeviceUpdating);
-        }
-        Ok(UpdateCheckResult::VersionCheckFailed) => {
-            // Preserve the durable phase check_and_update_system already restored (Ready/Pairing);
-            // only collapse to Idle for non-durable phases. Forcing Idle here would demote a device
-            // that just recovered from update_failed via this BLE retry.
-            app_state
-                .lifecycle
-                .set(ble_terminal_reset_phase(app_state.lifecycle.get()));
-            return Err(ble::BleStatus::VersionCheckFailed);
-        }
-        Ok(UpdateCheckResult::NoUpdateNeeded) => {} // Continue with normal flow
-        Err(e) => {
-            // This shouldn't happen with NonBlocking, but handle it anyway
-            eprintln!("MAIN: Error during update check: {e:#?}");
-            app_state
-                .lifecycle
-                .set(ble_terminal_reset_phase(app_state.lifecycle.get()));
-            return Err(ble::BleStatus::VersionCheckFailed);
-        }
-    }
-
-    // Reuse a persisted topic instead of re-fetching from controld on every BLE success.
-    // get_relayer_info() blocks up to 31s and the BLE handler awaits this callback before replying,
-    // so a Pairing/Ready device (which already has its topic) must return immediately rather than
-    // risk a slow/failed GetRelayerTopicID. Only first-time setup (no usable persisted topic) hits
-    // the network. See needs_relayer_topic_fetch.
-    let state_store = &app_state.state_store;
-    let persisted_topic = state_store.get(persistent_state::TOPIC_ID);
-    let topic_id = if needs_relayer_topic_fetch(persisted_topic.as_deref()) {
-        let fetched = match dbus_utils::get_relayer_info() {
-            Ok(info) => info,
-            Err(e) => {
-                eprintln!("BLE: can't get relayer data from controld: {e:#?}");
-                return Err(ble::BleStatus::ServerUnreachable);
-            }
-        };
-
-        // Save topic_id to disk BEFORE setting Pairing phase to maintain invariant.
-        // Fail the operation if save fails, so we don't create invalid state.
-        state_store.set(persistent_state::TOPIC_ID, &fetched);
-        if let Err(e) = state_store.save() {
-            eprintln!("BLE: Failed to save topic_id: {e:#?}");
-            return Err(ble::BleStatus::UnknownError);
-        }
-        fetched
-    } else {
-        // Unwrap is safe: needs_relayer_topic_fetch returned false only for Some(non-empty).
-        persisted_topic.expect("persisted topic present when fetch is skipped")
-    };
-
-    // Only transition to Pairing if not already Ready.
-    // Ready devices completing a BLE retry should stay Ready, not be demoted to Pairing.
-    // This preserves the durable Ready phase across update check cycles.
-    let current_phase = app_state.lifecycle.get();
-    if current_phase != SetupPhase::Ready {
-        // NOW safe to set Pairing phase (topic_id is guaranteed to exist on disk).
-        // Pairing invariant: phase can only be Pairing when topic_id exists.
-        app_state.lifecycle.set(SetupPhase::Pairing);
-        if let Err(e) = app_state.lifecycle.persist(state_store) {
-            eprintln!("BLE: Failed to persist Pairing phase: {e:#?}");
-            // Phase is set in memory but not persisted - acceptable since topic_id is saved.
-            // Next boot will restore from disk (Idle) and can re-fetch topic_id.
-        }
-    }
-
-    let app_state = app_state.clone();
-    let chromium = chromium.clone();
-    task::spawn(async move {
-        let _ = show_message(&chromium, &app_state, constant::SETUP_SUCCESSFULLY_MSG).await;
-    });
-    Ok(topic_id)
-}
-
-mod callbacks {
-    use super::{
-        AppState, Cdp, SetupPhase, UpdateGuard, WifiError, ble, constant, dbus_utils,
-        get_setup_phase, internet_setup_successfully_cb, pairing_confirmation_promotes_to_ready,
-        qr_switch_blocked_by_update_failed, show_factory_reset, show_message, show_qrcode,
-        show_webapp, wifi_utils,
-    };
-    use serde::Deserialize;
-    use std::sync::Arc;
-    use std::sync::atomic::Ordering;
-    use std::time::{Duration, Instant};
-    use tokio::task;
-
-    #[derive(Deserialize)]
-    struct BundledUploadLogsPayload {
-        api_key: String,
-        support_bundle_id: String,
-    }
-
-    fn parse_bundled_upload_logs_payload(
-        payload: &[u8],
-    ) -> serde_json::Result<BundledUploadLogsPayload> {
-        serde_json::from_slice::<BundledUploadLogsPayload>(payload)
-    }
-
-    pub fn create_bt_connected_cb(
-        app_state: Arc<AppState>,
-        chromium: Arc<Cdp>,
-    ) -> ble::BTConnectedCallback {
-        Some(Box::new(move || {
-            let app_state = app_state.clone();
-            let chromium = chromium.clone();
-
-            Box::pin(async move {
-                let should_show_welcome = {
-                    let page = app_state.page.lock().await;
-                    matches!(*page, super::Page::QRCode(_))
-                };
-                if should_show_welcome {
-                    let _ = show_message(&chromium, &app_state, constant::WELCOME_MSG).await;
-                }
-            })
-        }))
-    }
-
-    pub fn create_bt_disconnected_cb(
-        app_state: Arc<AppState>,
-        chromium: Arc<Cdp>,
-    ) -> ble::BTDisconnectedCallback {
-        Some(Box::new(move || {
-            let app_state = app_state.clone();
-            let chromium = chromium.clone();
-
-            Box::pin(async move {
-                // Reads the canonical `page`. During a blocking version check that page is the
-                // real prior surface (e.g. webapp), not the transient "Checking..." copy — the
-                // progress UI navigates transiently and never records `page`. So a disconnect
-                // mid-check now acts on the actual surface (keeping webapp/recovery instead of
-                // dropping an already-settled device to QR), matching steady-state behavior.
-                let should_go_qrcode = {
-                    let page = app_state.page.lock().await;
-                    !page.should_keep_on_bt_disconnect()
-                };
-                if should_go_qrcode {
-                    let _ = show_qrcode(&app_state, &chromium).await;
-                }
-            })
-        }))
-    }
-
-    pub fn create_connect_wifi_cb(
-        app_state: Arc<AppState>,
-        chromium: Arc<Cdp>,
-    ) -> ble::ConnectWifiCallback {
-        Box::new(move |ssid, pwd| {
-            let app_state = app_state.clone();
-            let chromium = chromium.clone();
-            let ssid = ssid.to_string();
-            let pwd = pwd.to_string();
-            Box::pin(async move {
-                let start_time = Instant::now();
-
-                // Acquire device ownership BEFORE any side effects, serializing this retry against
-                // OTA for its entire duration. A load-only precheck left a TOCTOU window: it could
-                // observe "no update", a concurrent task could then acquire ownership, and this path
-                // would still switch Wi-Fi via nmcli (disrupting an in-flight download) and clobber
-                // the setup_phase/UI mobile polls. Holding the guard from here through the update
-                // check (handed to internet_setup_successfully_cb) closes that window. RAII releases
-                // it on every early return below.
-                let guard = match UpdateGuard::try_acquire(&app_state.update_in_progress) {
-                    Some(guard) => guard,
-                    None => {
-                        println!("BLE: connect_wifi rejected, update in progress");
-                        return Err(ble::BleStatus::DeviceUpdating);
-                    }
-                };
-
-                // Clear and restore UpdateFailed phase on explicit WiFi retry.
-                // Restore the pre-failure durable phase if tracked, otherwise use Idle.
-                if app_state.lifecycle.get() == SetupPhase::UpdateFailed {
-                    use crate::persistent_state::PRE_FAILURE_PHASE;
-
-                    let restored_phase =
-                        if let Some(phase_str) = app_state.state_store.get(PRE_FAILURE_PHASE) {
-                            let phase = SetupPhase::from_str(&phase_str);
-                            println!(
-                                "BLE: Restoring pre-failure phase '{}' on WiFi retry",
-                                phase.as_str()
-                            );
-                            phase
-                        } else {
-                            println!(
-                                "BLE: Clearing UpdateFailed to Idle (no pre-failure phase tracked)"
-                            );
-                            SetupPhase::Idle
-                        };
-
-                    app_state.lifecycle.set(restored_phase);
-
-                    // Clear the pre-failure tracking
-                    app_state.state_store.set(PRE_FAILURE_PHASE, "");
-
-                    // Abort the retry if the cleared UpdateFailed phase cannot be persisted.
-                    // Proceeding would leave setup_phase=update_failed on disk while the
-                    // in-memory phase says otherwise, so a crash/reboot mid-retry would
-                    // resurrect UpdateFailed. Mirrors the D-Bus do_system_update guard.
-                    if let Err(e) = app_state.lifecycle.persist(&app_state.state_store) {
-                        eprintln!(
-                            "BLE: Failed to persist phase restoration, aborting retry: {e:#?}"
-                        );
-                        return Err(ble::BleStatus::UnknownError);
-                    }
-
-                    // Re-save durable phase to PRE_FAILURE_PHASE before entering transient state.
-                    // This ensures the phase survives the update check cycle and can be restored
-                    // if the retry fails again. Without this, a second failure would have no
-                    // durable phase to restore from, permanently losing Ready/Pairing state.
-                    if restored_phase.is_durable() && restored_phase != SetupPhase::UpdateFailed {
-                        app_state
-                            .state_store
-                            .set(PRE_FAILURE_PHASE, restored_phase.as_str());
-                        if let Err(e) = app_state.state_store.save() {
-                            eprintln!(
-                                "BLE: Failed to re-save PRE_FAILURE_PHASE before retry: {e:#?}"
-                            );
-                        }
-                    }
-                }
-
-                // Capture the durable phase (Ready/Pairing) this device had before the BLE setup.
-                // The mandatory update check inside internet_setup_successfully_cb snapshots the
-                // LIVE phase as its no-update restore target. WifiConnecting is transient, so if it
-                // were still live at that point a no-update result would collapse to Idle and the
-                // setup-success path would then demote a previously Ready device to Pairing. We
-                // restore this captured phase just before the check (see below).
-                let phase_before_wifi_connect = app_state.lifecycle.get();
-
-                app_state.lifecycle.set(SetupPhase::WifiConnecting);
-                // Show message
-                let connecting_msg = format!("{}{}", constant::WIFI_CONNECTING_MSG_PREFIX, ssid);
-                let _ = show_message(&chromium, &app_state, &connecting_msg).await;
-
-                // Disable auto proceed since users want to setup another wifi
-                // Instead of fixing the current internet (if there is any)
-                app_state.auto_proceed.store(false, Ordering::Release);
-
-                // Connect to wifi & return early if failed
-                if let Err(e) = wifi_utils::connect(&ssid, &pwd).await {
-                    eprintln!(
-                        "MAIN: Failed to connect to wifi \"{ssid}\" in {:?} ms: {e: }",
-                        start_time.elapsed().as_millis()
-                    );
-                    app_state.lifecycle.set(SetupPhase::Idle);
-                    // Tell user that the wifi connection failed
-                    task::spawn(async move {
-                        let _ = show_message(
-                            &chromium,
-                            &app_state,
-                            constant::WIFI_FAILED_TO_CONNECT_MSG,
-                        )
-                        .await;
-                    });
-                    // This is a bit of a hack to detect wrong password
-                    // But the command doesn't provide a reliable way to detect this
-                    let err_code = match &e {
-                        WifiError::NmcliFailure { stderr, .. } if stderr.contains("password") => {
-                            ble::BleStatus::WrongWifiPassword
-                        }
-                        _ => ble::BleStatus::UnknownError,
-                    };
-                    return Err(err_code);
-                }
-
-                // Wait for internet connectivity (up to 6 seconds)
-                // WiFi may be connected but internet routing can take a moment
-                app_state
-                    .internet
-                    .wait_until_online(
-                        Duration::from_millis(constant::WIFI_INTERNET_CHECK_INTERVAL),
-                        Some(Duration::from_millis(constant::WIFI_INTERNET_WAIT_TIMEOUT)),
-                    )
-                    .await;
-
-                // Return early if there is still no internet after waiting
-                if !app_state.internet.is_online(true).await {
-                    app_state.lifecycle.set(SetupPhase::Idle);
-                    task::spawn(async move {
-                        let _ = show_message(
-                            &chromium,
-                            &app_state,
-                            constant::INTERNET_FAILED_TO_CONNECT_MSG,
-                        )
-                        .await;
-                    });
-                    return Err(ble::BleStatus::NoInternet);
-                }
-
-                // Restore the pre-connect durable phase so the mandatory update check's no-update
-                // restoration keeps Ready/Pairing instead of resetting the transient WifiConnecting
-                // to Idle (which would demote a previously Ready device to Pairing in the
-                // setup-success path). Idle/Pairing devices are unchanged: a first-time Idle device
-                // still flows to topic_id fetch -> Pairing below.
-                app_state.lifecycle.set(phase_before_wifi_connect);
-                internet_setup_successfully_cb(&app_state, &chromium, guard).await
-            })
-        })
-    }
-
-    pub fn create_keep_wifi_cb(
-        app_state: Arc<AppState>,
-        chromium: Arc<Cdp>,
-    ) -> ble::KeepWifiCallback {
-        Box::new(move || {
-            let app_state = app_state.clone();
-            let chromium = chromium.clone();
-            Box::pin(async move {
-                // Acquire device ownership before any phase/UI side effects so the whole retry is
-                // serialized against OTA (mirrors connect_wifi). A load-only precheck left a TOCTOU
-                // window where a concurrent task could acquire ownership after the check while this
-                // path still cleared/restored the persisted UpdateFailed latch. RAII releases the
-                // guard on every early return below.
-                let guard = match UpdateGuard::try_acquire(&app_state.update_in_progress) {
-                    Some(guard) => guard,
-                    None => {
-                        println!("BLE: keep_wifi rejected, update in progress");
-                        return Err(ble::BleStatus::DeviceUpdating);
-                    }
-                };
-
-                if !app_state.internet.is_online(true).await {
-                    return Err(ble::BleStatus::WifiRequired);
-                }
-
-                // Clear and restore UpdateFailed phase on explicit keep_wifi retry.
-                // Restore the pre-failure durable phase if tracked, otherwise use Idle.
-                if app_state.lifecycle.get() == SetupPhase::UpdateFailed {
-                    use crate::persistent_state::PRE_FAILURE_PHASE;
-
-                    let restored_phase =
-                        if let Some(phase_str) = app_state.state_store.get(PRE_FAILURE_PHASE) {
-                            let phase = SetupPhase::from_str(&phase_str);
-                            println!(
-                                "BLE: Restoring pre-failure phase '{}' on keep_wifi retry",
-                                phase.as_str()
-                            );
-                            phase
-                        } else {
-                            println!(
-                                "BLE: Clearing UpdateFailed to Idle (no pre-failure phase tracked)"
-                            );
-                            SetupPhase::Idle
-                        };
-
-                    app_state.lifecycle.set(restored_phase);
-
-                    // Clear the pre-failure tracking
-                    app_state.state_store.set(PRE_FAILURE_PHASE, "");
-
-                    // Abort the retry if the cleared UpdateFailed phase cannot be persisted.
-                    // Proceeding would leave setup_phase=update_failed on disk while the
-                    // in-memory phase says otherwise, so a crash/reboot mid-retry would
-                    // resurrect UpdateFailed. Mirrors the D-Bus do_system_update guard.
-                    if let Err(e) = app_state.lifecycle.persist(&app_state.state_store) {
-                        eprintln!(
-                            "BLE: Failed to persist phase restoration, aborting retry: {e:#?}"
-                        );
-                        return Err(ble::BleStatus::UnknownError);
-                    }
-
-                    // Re-save durable phase to PRE_FAILURE_PHASE before entering transient state.
-                    // This ensures the phase survives the update check cycle and can be restored
-                    // if the retry fails again. Without this, a second failure would have no
-                    // durable phase to restore from, permanently losing Ready/Pairing state.
-                    if restored_phase.is_durable() && restored_phase != SetupPhase::UpdateFailed {
-                        app_state
-                            .state_store
-                            .set(PRE_FAILURE_PHASE, restored_phase.as_str());
-                        if let Err(e) = app_state.state_store.save() {
-                            eprintln!(
-                                "BLE: Failed to re-save PRE_FAILURE_PHASE before retry: {e:#?}"
-                            );
-                        }
-                    }
-                }
-
-                internet_setup_successfully_cb(&app_state, &chromium, guard).await
-            })
-        })
-    }
-
-    pub fn create_get_info_cb(app_state: Arc<AppState>) -> ble::GetInfoCallback {
-        Some(Box::new(move || {
-            let phase = get_setup_phase(&app_state);
-            if matches!(phase.as_str(), "updating" | "checking_version") {
-                // Trigger async refresh to avoid blocking BLE response on D-Bus timeout.
-                // Fresh value will be available on next poll (mobile polls every few seconds).
-                let _ = app_state.internet.trigger_refresh_async();
-            }
-            vec![super::build_device_info(&app_state)]
-        }))
-    }
-
-    pub fn create_qrcode_switch_cb(
-        app_state: Arc<AppState>,
-        chromium: Arc<Cdp>,
-    ) -> dbus_utils::ListenCallback {
-        Box::new(move |msg| {
-            println!("MAIN: QR switch callback received");
-            let chromium = chromium.clone();
-            let app_state = app_state.clone();
-            let mut qrcode_requested = false;
-            match msg.read1::<bool>() {
-                Ok(true) => qrcode_requested = true,
-                Err(e) => println!("MAIN: Error reading message: {e: }"),
-                _ => {}
-            }
-            println!("MAIN: QR switch -> qrcode_requested={qrcode_requested}");
-            tokio::runtime::Handle::current().block_on(async move {
-                // A `show_pairing_qr_code(false)` pairing confirmation is a durable, one-shot signal:
-                // controld ACKs it after this callback returns and does not re-emit. Record the
-                // Pairing -> Ready transition BEFORE the navigation ownership gate so it is never
-                // lost when an OTA owns the device. (Regression from afd9ee4, which gated the whole
-                // switch and dropped this persist; the load-only check was kept by a898a72.) The
-                // transition is also valid when an in-flight OTA op transiently masks Pairing with a
-                // non-durable phase while a topic_id is already persisted; check_and_update_system's
-                // restore then refuses to demote this Ready back to Pairing (phase_after_inflight_check).
-                if !qrcode_requested {
-                    let has_topic = app_state
-                        .state_store
-                        .get(crate::persistent_state::TOPIC_ID)
-                        .is_some_and(|t| !t.is_empty());
-                    if pairing_confirmation_promotes_to_ready(app_state.lifecycle.get(), has_topic)
-                    {
-                        println!("MAIN: QR switch -> transitioning to Ready phase");
-                        app_state.lifecycle.set(SetupPhase::Ready);
-                        if let Err(e) = app_state.lifecycle.persist(&app_state.state_store) {
-                            eprintln!("MAIN: Error persisting Ready phase: {e:#?}");
-                        }
-                    }
-                }
-
-                // Navigation must not clobber a screen owned by an active OTA. Acquire ownership for
-                // the navigation only; if an OTA owns the device, skip painting now — the durable
-                // transition above is already recorded and the post-update reboot (or a later
-                // canonical-page restore) repaints the correct surface. Held until this block
-                // returns (RAII), serializing the navigation against a concurrent update check.
-                let _guard = match UpdateGuard::try_acquire(&app_state.update_in_progress) {
-                    Some(guard) => guard,
-                    None => {
-                        println!("MAIN: QR switch navigation skipped, update in progress");
-                        return;
-                    }
-                };
-
-                // Stay on the permanent-failure surface until an explicit retry clears the latch.
-                // handle_permanent_update_failure releases the update guard but keeps setup_phase ==
-                // update_failed, so without this a QR/webapp switch from controld would navigate
-                // Chromium off the failure message while the device still reports update_failed.
-                // Only BLE connect_wifi/keep_wifi or D-Bus system_update clear the latch and move on.
-                if qr_switch_blocked_by_update_failed(app_state.lifecycle.get()) {
-                    println!(
-                        "MAIN: QR switch navigation skipped, update_failed latched (awaiting retry)"
-                    );
-                    return;
-                }
-
-                if qrcode_requested {
-                    let _ = show_qrcode(&app_state, &chromium).await;
-                } else {
-                    let _ = show_webapp(&app_state, &chromium).await;
-                }
-            });
-        })
-    }
-
-    async fn do_factory_reset(chromium: &Arc<Cdp>, app_state: &Arc<AppState>) {
-        // Show factory reset page
-        let _ = show_factory_reset(chromium, app_state).await;
-        // Execute factory reset
-        if let Err(e) = super::system::factory_reset().await {
-            eprintln!("MAIN: Failed to execute factory reset: {e:#?}");
-        }
-    }
-
-    async fn do_system_update(chromium: &Arc<Cdp>, app_state: &Arc<AppState>) {
-        // Acquire device ownership BEFORE any connectivity probe, no-internet navigation, or
-        // UpdateFailed latch clear/restore. A load-only precheck left a TOCTOU window where a
-        // concurrent task could acquire ownership after the check while this path still mutated the
-        // persisted latch and UI. Holding the guard from here through check_and_update_system (which
-        // receives it) serializes the whole manual retry against OTA. RAII releases it on return.
-        let guard = match UpdateGuard::try_acquire(&app_state.update_in_progress) {
-            Some(guard) => guard,
-            None => {
-                println!("MAIN: system_update ignored, update already in progress");
-                return;
-            }
-        };
-
-        // Check internet connectivity first before clearing latch
-        if !app_state.internet.is_online(true).await {
-            eprintln!("MAIN: System update requested but no internet connection");
-            let _ = show_message(
-                chromium,
-                app_state,
-                constant::INTERNET_FAILED_TO_CONNECT_MSG,
-            )
-            .await;
-            return;
-        }
-
-        // Clear and restore UpdateFailed phase on explicit D-Bus manual retry.
-        // Restore the pre-failure durable phase if tracked, otherwise use Idle.
-        if app_state.lifecycle.get() == SetupPhase::UpdateFailed {
-            use crate::persistent_state::PRE_FAILURE_PHASE;
-
-            let restored_phase =
-                if let Some(phase_str) = app_state.state_store.get(PRE_FAILURE_PHASE) {
-                    let phase = SetupPhase::from_str(&phase_str);
-                    println!(
-                        "DBUS: Restoring pre-failure phase '{}' on manual system update",
-                        phase.as_str()
-                    );
-                    phase
-                } else {
-                    println!("DBUS: Clearing UpdateFailed to Idle (no pre-failure phase tracked)");
-                    SetupPhase::Idle
-                };
-
-            app_state.lifecycle.set(restored_phase);
-
-            // Clear the pre-failure tracking
-            app_state.state_store.set(PRE_FAILURE_PHASE, "");
-
-            if let Err(e) = app_state.lifecycle.persist(&app_state.state_store) {
-                eprintln!("DBUS: Failed to persist phase restoration: {e:#?}");
-                return;
-            }
-        }
-
-        // Use Available mode for user-triggered updates (check for any newer version)
-        // Use Blocking execution since D-Bus callback runs in a spawned task anyway
-        let result = super::check_and_update_system(
-            app_state,
-            chromium,
-            super::UpdateMode::Available,
-            super::UpdateExecution::Blocking,
-            guard,
-        )
-        .await;
-
-        match super::available_system_update_follow_up(&result) {
-            super::AvailableSystemUpdateFollowUp::RestoreCurrentPage => {
-                // The blocking version-check progress UI repaints Chromium with the transient
-                // "Checking for updates..." screen and intentionally does NOT record
-                // app_state.page (see navigate_transient_message). So after a no-update result
-                // the TV is left on that transient screen while app_state.page still holds the
-                // real canonical surface. We must actively re-show that canonical page or the
-                // device stays stuck on "Checking for updates...". The decision is factored into
-                // restore_page_target so it can be unit-tested without driving CDP.
-                let current_page = { app_state.page.lock().await.clone() };
-                let target = super::restore_page_target(&current_page, app_state.lifecycle.get());
-                let nav_result = match target {
-                    super::RestorePageTarget::Webapp => {
-                        super::show_webapp(app_state, chromium).await
-                    }
-                    super::RestorePageTarget::Qrcode => {
-                        super::show_qrcode(app_state, chromium).await
-                    }
-                    super::RestorePageTarget::Message(msg) => {
-                        super::show_message(chromium, app_state, &msg).await
-                    }
-                    super::RestorePageTarget::FactoryReset => {
-                        super::show_factory_reset(chromium, app_state).await
-                    }
-                    super::RestorePageTarget::NoChange => Ok(()),
-                };
-                if let Err(e) = nav_result {
-                    eprintln!(
-                        "MAIN: Failed to restore canonical page after D-Bus no-update: {e:#?}"
-                    );
-                }
-            }
-            super::AvailableSystemUpdateFollowUp::LogFailure => {
-                eprintln!("MAIN: System update failed: {result:#?}");
-            }
-            super::AvailableSystemUpdateFollowUp::NoOp => {}
-        }
-    }
-
-    pub fn create_factory_reset_cb(
-        app_state: Arc<AppState>,
-        chromium: Arc<Cdp>,
-    ) -> ble::FactoryResetCallback {
-        Some(Box::new(move || {
-            let chromium = chromium.clone();
-            let app_state = app_state.clone();
-            Box::pin(async move {
-                do_factory_reset(&chromium, &app_state).await;
-            })
-        }))
-    }
-
-    pub fn create_factory_reset_dbus_cb(
-        app_state: Arc<AppState>,
-        chromium: Arc<Cdp>,
-    ) -> dbus_utils::ListenCallback {
-        Box::new(move |_msg| {
-            println!("MAIN: Factory reset DBus callback received");
-            let chromium = chromium.clone();
-            let app_state = app_state.clone();
-            task::spawn(async move {
-                do_factory_reset(&chromium, &app_state).await;
-            });
-        })
-    }
-
-    pub fn create_system_update_dbus_cb(
-        app_state: Arc<AppState>,
-        chromium: Arc<Cdp>,
-    ) -> dbus_utils::ListenCallback {
-        Box::new(move |_msg| {
-            println!("MAIN: System update DBus callback received");
-            let chromium = chromium.clone();
-            let app_state = app_state.clone();
-            task::spawn(async move {
-                do_system_update(&chromium, &app_state).await;
-            });
-        })
-    }
-
-    /// Core log upload logic - zips logs folder and uploads via v2 API
-    async fn do_upload_logs(
-        device_id: &str,
-        api_key: &str,
-        source: &str,
-        branch: &str,
-        version: &str,
-        support_bundle_id: Option<&str>,
-    ) {
-        println!("MAIN: Uploading logs (source: {source})");
-
-        if let Err(e) = super::log_uploader::submit_logs(
-            device_id,
-            api_key,
-            source,
-            branch,
-            version,
-            support_bundle_id,
-        )
-        .await
-        {
-            eprintln!("MAIN: Failed to submit logs: error code {e}");
-        } else {
-            println!("MAIN: Logs submitted successfully");
-        }
-    }
-
-    pub fn create_submit_logs_cb(app_state: Arc<AppState>) -> ble::SubmitLogsCallback {
-        Box::new(move |_user_id, api_key, _title, support_bundle_id| {
-            let api_key = api_key.to_string();
-            let support_bundle_id = support_bundle_id.map(str::to_string);
-            let device_id = app_state.device_id.clone();
-            let branch = app_state.branch.clone();
-            let version = app_state.current_version.clone();
-            Box::pin(async move {
-                do_upload_logs(
-                    &device_id,
-                    &api_key,
-                    "ble",
-                    &branch,
-                    &version,
-                    support_bundle_id.as_deref(),
-                )
-                .await;
-            })
-        })
-    }
-
-    pub fn create_upload_logs_dbus_cb(app_state: Arc<AppState>) -> dbus_utils::ListenCallback {
-        Box::new(move |msg| {
-            println!("MAIN: Upload logs DBus callback received");
-            // Read parameters: user_id, api_key, title (title unused in v2 API)
-            let (_user_id, api_key, _title): (String, String, String) =
-                match msg.read3::<String, String, String>() {
-                    Ok(params) => params,
-                    Err(e) => {
-                        eprintln!("MAIN: Failed to read upload logs parameters: {e:#?}");
-                        return;
-                    }
-                };
-
-            let device_id = app_state.device_id.clone();
-            let branch = app_state.branch.clone();
-            let version = app_state.current_version.clone();
-            task::spawn(async move {
-                do_upload_logs(&device_id, &api_key, "dbus", &branch, &version, None).await;
-            });
-        })
-    }
-
-    pub fn create_upload_logs_with_bundle_dbus_cb(
-        app_state: Arc<AppState>,
-    ) -> dbus_utils::ListenCallback {
-        Box::new(move |msg| {
-            println!("MAIN: Upload logs with bundle DBus callback received");
-            let payload = match msg.read1::<Vec<u8>>() {
-                Ok(payload) => payload,
-                Err(e) => {
-                    eprintln!("MAIN: Failed to read bundled upload logs payload: {e:#?}");
-                    return;
-                }
-            };
-            let payload = match parse_bundled_upload_logs_payload(&payload) {
-                Ok(payload) => payload,
-                Err(e) => {
-                    eprintln!("MAIN: Failed to parse bundled upload logs payload: {e:#?}");
-                    return;
-                }
-            };
-
-            let device_id = app_state.device_id.clone();
-            let branch = app_state.branch.clone();
-            let version = app_state.current_version.clone();
-            task::spawn(async move {
-                do_upload_logs(
-                    &device_id,
-                    &payload.api_key,
-                    "dbus",
-                    &branch,
-                    &version,
-                    Some(&payload.support_bundle_id),
-                )
-                .await;
-            });
-        })
-    }
-
-    #[cfg(test)]
-    mod tests {
-        use super::*;
-
-        #[test]
-        fn parse_bundled_upload_logs_payload_accepts_controld_schema() {
-            let payload = br#"{
-                "user_id": "ignored-user",
-                "api_key": "test-api-key",
-                "title": "ignored-title",
-                "support_bundle_id": "bundle-123"
-            }"#;
-
-            let parsed =
-                parse_bundled_upload_logs_payload(payload).expect("parse bundled upload payload");
-
-            assert_eq!(parsed.api_key, "test-api-key");
-            assert_eq!(parsed.support_bundle_id, "bundle-123");
-        }
-    }
-}
-
-// device_info is <device_id>|<topic_id>|<internet>|<branch>|<version>|<setup_phase>
-fn build_device_info(app_state: &Arc<AppState>) -> String {
-    let device_id = app_state.device_id.clone();
-    let topic_id = app_state
-        .state_store
-        .get(persistent_state::TOPIC_ID)
-        .unwrap_or_default();
-    let has_internet = if app_state.internet.is_online_cached() {
-        "true"
-    } else {
-        "false"
-    };
-    let branch = app_state.branch.clone().replace('/', "%2F");
-    let version = app_state.current_version.clone();
-    let setup_phase = get_setup_phase(app_state);
-
-    format!("{device_id}|{topic_id}|{has_internet}|{branch}|{version}|{setup_phase}")
-}
-
-// The url format is like this
-// url?step=qr&device_info=<device_info>&version=<version>&device_id=<device_id>
-// Note: we need extra device_id and version for the page to display at the bottom
-fn build_qrcode_url(app_state: &Arc<AppState>) -> String {
-    let device_info = build_device_info(app_state);
-    let device_id = app_state.device_id.clone();
-    let version = app_state.current_version.clone();
-
-    format!(
-        "{}&device_info={device_info}&version={version}&device_id={device_id}",
-        constant::QRCODE_URL_PREFIX
-    )
 }
 
 async fn wait_for_shutdown() {
@@ -1547,1011 +139,36 @@ async fn wait_for_shutdown() {
     }
 }
 
-async fn show_qrcode(app_state: &Arc<AppState>, chrome: &Arc<Cdp>) -> Result<()> {
-    // Normal QR navigation resets non-failure transient phases to idle (WifiConnecting,
-    // CheckingVersion, Updating), while preserving UpdateFailed, Pairing, and Ready phases.
-    // This ensures the device returns to a clean setup state when showing QR outside of
-    // explicit failure recovery flows.
-    let current = app_state.lifecycle.get();
-    if matches!(
-        current,
-        SetupPhase::WifiConnecting | SetupPhase::CheckingVersion | SetupPhase::Updating
-    ) {
-        app_state.lifecycle.set(SetupPhase::Idle);
-    }
-
-    let qrcode_url = build_qrcode_url(app_state);
-    // QRCode url is dynamically built
-    // So we always navigate to make sure the url is correct
-    let mut page = app_state.page.lock().await;
-    chrome
-        .navigate(&qrcode_url)
-        .await
-        .with_context(|| format!("navigating to {qrcode_url}"))?;
-    println!("MAIN: Navigated to {qrcode_url}");
-    *page = Page::QRCode(unix_s());
-    Ok(())
-}
-
-async fn show_reflashing_qrcode(
-    app_state: &Arc<AppState>,
-    chrome: &Arc<Cdp>,
-    flashing_guide_url: &str,
-    current_version: &str,
-    latest_version: &str,
-    min_upgradeable_version: &str,
-) -> Result<()> {
-    // Build message with version information
-    let message = format!(
-        "We've moved too far ahead for this version to catch up. Your FF1 is too far behind to auto-upgrade. Current version: {current_version} Latest version: {latest_version} Minimum upgradeable version: {min_upgradeable_version}. Scan the code above for step-by-step reflashing instructions, or contact us for help. support@feralfile.com"
-    );
-
-    // Build URL with QR code step and flashing guide as the QR content
-    let qrcode_url = format!(
-        "{}&qr_content={}&message={}",
-        constant::QRCODE_URL_PREFIX,
-        urlencoding::encode(flashing_guide_url),
-        urlencoding::encode(&message)
-    );
-
-    let mut page = app_state.page.lock().await;
-    chrome
-        .navigate(&qrcode_url)
-        .await
-        .with_context(|| format!("navigating to {qrcode_url}"))?;
-    println!("MAIN: Navigated to reflashing QR code: {qrcode_url}");
-    *page = Page::ReflashingRequired(unix_s(), message);
-    Ok(())
-}
-
-/// Handles the main startup flow once the device has a working internet connection.
-///
-/// When this is called:
-/// - The app state and CDP connection have already been initialized.
-/// - The caller has determined that the device currently has internet access.
-///
-/// What it does:
-/// - Checks whether the running firmware/software is too old to auto-upgrade and, if so,
-///   shows a reflashing QR code or a fallback message and stops further processing.
-/// - If the device can be upgraded, checks whether an update is required and either
-///   drives the updater flow or continues with normal startup.
-/// - If no update is in progress, decides whether to show the web app or the pairing
-///   QR code based on the presence of a cached topic ID and whether the device is in
-///   qemu mode.
-///
-/// Notes:
-/// - Any early return from this function (for example, when an update is required or
-///   the device is too old) is intentional and means the usual "show art or QR" flow
-///   should not continue.
-async fn on_startup_with_internet(app_state: Arc<AppState>, chrome: Arc<Cdp>) -> Result<()> {
-    // If UpdateFailed phase is set, skip automatic update check on startup.
-    // Show the failure message (different from fresh failure) since this is reboot recovery.
-    // Mobile app will see update_failed via device_info polling and can trigger explicit retry.
-    if startup_requires_update_failed_recovery(app_state.lifecycle.get()) {
-        println!("MAIN: Skipping startup update check - UpdateFailed phase is set");
-        show_update_failed_recovery(&app_state, &chrome).await?;
-        return Ok(());
-    }
-
-    // Check and update system using Required mode (only mandatory updates on startup).
-    // This runs for ALL phases except UpdateFailed, maintaining consistency:
-    // - First-time setup (Idle): checks before fetching topic_id → Pairing
-    // - Reboot with Pairing: checks again to ensure no new mandatory updates
-    // - Reboot with Ready: checks to keep device up-to-date
-    // Use Blocking execution since we can wait for completion during startup.
-    //
-    // Acquire device ownership before the check. In the rare case an OTA already owns the device
-    // (e.g. a BLE retry started one during the offline->online transition), skip the startup check
-    // entirely and stay alive; the owner drives the UI. This matches the old behavior where the
-    // self-acquire failed and returned UpdateInProgress (-> Halt -> Ok), just without the redundant
-    // result hop.
-    let guard = match UpdateGuard::try_acquire(&app_state.update_in_progress) {
-        Some(guard) => guard,
-        None => {
-            println!("MAIN: startup update check skipped, update already in progress");
-            return Ok(());
-        }
-    };
-    let check_result = check_and_update_system(
-        &app_state,
-        &chrome,
-        UpdateMode::Required,
-        UpdateExecution::Blocking,
-        guard,
-    )
-    .await?;
-    match startup_update_check_outcome(check_result) {
-        // Too old, updating, already running, or a failed version check. The core check already
-        // drove the canonical UI and restored the phase; stay alive and stop here so we do not
-        // overwrite that surface or exit the daemon. See startup_update_check_outcome for why a
-        // failed check must NOT be fatal.
-        StartupUpdateOutcome::Halt => return Ok(()),
-        StartupUpdateOutcome::Continue => {} // No update needed: continue with normal flow.
-    }
-
-    // No update needed. Show UI based on current phase.
-    // If we don't have a topic_id yet, try to get one and transition to Pairing.
-    let state_store = &app_state.state_store;
-    let current_phase = app_state.lifecycle.get();
-
-    // If still in Idle and don't have topic_id, try to get it
-    if current_phase == SetupPhase::Idle && state_store.get(persistent_state::TOPIC_ID).is_none() {
-        match dbus_utils::get_relayer_info() {
-            Ok(topic_id) => {
-                // Save topic_id FIRST before setting Pairing phase
-                state_store.set(persistent_state::TOPIC_ID, &topic_id);
-                if let Err(e) = state_store.save() {
-                    eprintln!("MAIN: Failed to save topic_id: {e:#?}");
-                    // Don't transition to Pairing if save failed - keep Idle
-                    // Device will retry on next boot or BLE flow
-                } else {
-                    // Topic_id saved successfully, now safe to transition to Pairing
-                    app_state.lifecycle.set(SetupPhase::Pairing);
-                    if let Err(e) = app_state.lifecycle.persist(state_store) {
-                        eprintln!("MAIN: Error persisting Pairing phase: {e:#?}");
-                        // Phase set in memory but not persisted - acceptable since topic_id is saved
-                    }
-                }
-            }
-            Err(e) => {
-                eprintln!(
-                    "MAIN: startup_with_internet: can't get relayer data from controld: {e:#?}"
-                );
-            }
-        }
-    }
-
-    // Show UI based on phase
-    let phase = app_state.lifecycle.get();
-    println!("MAIN: startup_with_internet: phase={}", phase.as_str());
-
-    match phase {
-        SetupPhase::Ready => show_webapp(&app_state, &chrome).await,
-        _ => show_qrcode(&app_state, &chrome).await,
-    }
-}
-
-async fn update(app_state: Arc<AppState>, chrome: Arc<Cdp>) -> Result<()> {
-    app_state.lifecycle.set(SetupPhase::Updating);
-
-    let latest_version = updater::latest_version().await.unwrap_or_else(|e| {
-        eprintln!("MAIN: latest_version for update banner: {e:#?}");
-        "Unknown".to_string()
-    });
-    let base_msg = format!("{} {}", &constant::UPDATING_MSG_PREFIX, latest_version);
-    let default_subtext = constant::UPDATING_MSG_SUBTEXT;
-
-    let mut attempt = 0u8;
-    loop {
-        attempt += 1;
-
-        if attempt == 1 {
-            let _ = show_system_upgrade(
-                &chrome,
-                &app_state,
-                &format!("{base_msg}&subtext={default_subtext}"),
-            )
-            .await;
-        }
-
-        match attempt_update(&app_state, &chrome, &base_msg).await {
-            Ok(()) => {
-                // Update succeeded. Reset to Idle; if device was already paired,
-                // phase will be restored to Ready on next boot from persistent state.
-                //
-                // Clear the stale recovery marker. PRE_FAILURE_PHASE was written before
-                // entering the transient update states, but is only overwritten on a
-                // subsequent check when the prior phase is durable. Leaving it set would
-                // let a later failure that starts from a non-durable phase restore a
-                // phantom Ready/Pairing during update_failed recovery.
-                use crate::persistent_state::PRE_FAILURE_PHASE;
-                app_state.state_store.set(PRE_FAILURE_PHASE, "");
-                app_state.lifecycle.set(SetupPhase::Idle);
-                if let Err(e) = app_state.state_store.save() {
-                    eprintln!("MAIN: Failed to clear PRE_FAILURE_PHASE on update success: {e:#?}");
-                }
-                return Ok(());
-            }
-            Err(e) => {
-                match e.kind {
-                    UpdateErrorType::Transient if attempt < MAX_UPDATE_RETRIES => {
-                        eprintln!(
-                            "MAIN: Transient update error (attempt {attempt}/{MAX_UPDATE_RETRIES}): {e:#?}"
-                        );
-                        let retry_msg = format!(
-                            "Retrying update (attempt {}/{MAX_UPDATE_RETRIES})",
-                            attempt + 1
-                        );
-                        let _ = show_system_upgrade(
-                            &chrome,
-                            &app_state,
-                            &format!("{base_msg}&subtext={retry_msg}"),
-                        )
-                        .await;
-                        // Exponential backoff: 2^attempt seconds (2s, 4s, 8s, ...)
-                        let backoff_seconds = 1u64 << attempt; // 2^attempt
-                        time::sleep(Duration::from_secs(backoff_seconds)).await;
-                        continue;
-                    }
-                    UpdateErrorType::Permanent | UpdateErrorType::Transient => {
-                        handle_permanent_update_failure(&app_state, &chrome, &e.source).await;
-                        return Err(e.source.context("update failed permanently"));
-                    }
-                }
-            }
-        }
-    }
-}
-
-async fn attempt_update(
-    app_state: &Arc<AppState>,
-    chrome: &Arc<Cdp>,
-    base_msg: &str,
-) -> Result<(), UpdateAttemptError> {
-    let mut rx = updater::spawn_updater().map_err(|e| UpdateAttemptError {
-        kind: classify_updater_message(&e.to_string()),
-        source: e,
-    })?;
-    let mut received_progress = false;
-
-    while let Some(res) = rx.recv().await {
-        match res {
-            Ok(msg) => {
-                received_progress = true;
-                let _ =
-                    show_system_upgrade(chrome, app_state, &format!("{base_msg}&subtext={msg}"))
-                        .await;
-            }
-            Err(e) => {
-                let _ = show_system_upgrade(chrome, app_state, &format!("{base_msg}&subtext={e}"))
-                    .await;
-                let kind = classify_updater_message(&e.to_string());
-                return Err(UpdateAttemptError {
-                    source: e.context("update process failed"),
-                    kind,
-                });
-            }
-        }
-    }
-
-    // Channel closed without receiving any progress - updater failed to start.
-    // This happens when watchdog stop fails, systemd start fails, or log file not found.
-    if !received_progress {
-        let message = "updater closed channel without sending progress";
-        return Err(UpdateAttemptError {
-            kind: classify_updater_message(message),
-            source: anyhow::anyhow!(message),
-        });
-    }
-
-    Ok(())
-}
-
-/// Show a clear failure message and keep it visible while `setup_phase=update_failed`.
-/// Mobile app can trigger retry via `keep_wifi` or `connect_wifi` BLE commands.
-/// Uses fresh failure message since this is called immediately after update failure.
-async fn handle_permanent_update_failure(
-    app_state: &Arc<AppState>,
-    chrome: &Arc<Cdp>,
-    error: &anyhow::Error,
-) {
-    eprintln!("MAIN: Permanent update failure: {error:#?}");
-
-    // PRE_FAILURE_PHASE was already stored in check_and_update_system before entering
-    // CheckingVersion, capturing the original Ready/Pairing phase before transient states.
-
-    app_state.lifecycle.set(SetupPhase::UpdateFailed);
-    if let Err(e) = app_state.lifecycle.persist(&app_state.state_store) {
-        eprintln!("MAIN: Failed to persist UpdateFailed phase: {e:#?}");
-        eprintln!("MAIN: Phase is set in memory; mobile sees update_failed until restart");
-    }
-    // Show fresh failure message - user can try restarting or retry via phone
-    let _ = show_system_upgrade(chrome, app_state, UPDATE_FAILED_FRESH_MSG).await;
-}
-
-/// Result of checking and potentially executing a system update.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum UpdateCheckResult {
-    /// Device is too old to auto-upgrade; reflashing is required.
-    TooOldToUpgrade,
-    /// An update was started (device will reboot after completion).
-    UpdateStarted,
-    /// No update is needed/available for the given mode.
-    NoUpdateNeeded,
-    /// Failed to check version information from the server.
-    VersionCheckFailed,
-    /// Update already in progress; concurrent attempt was skipped.
-    UpdateInProgress,
-}
-
-/// How `on_startup_with_internet` proceeds after the mandatory startup update check.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum StartupUpdateOutcome {
-    /// Stop the startup flow early and leave the daemon running in steady state. The core check
-    /// already drove the canonical UI (reflashing screen, updater-owned progress, or failure copy)
-    /// and restored the correct phase, so startup must NOT continue to topic_id/QR/webapp logic.
-    Halt,
-    /// Continue to topic_id fetch and surface selection (the normal no-update path).
-    Continue,
-}
-
-/// Map a mandatory startup update-check result to the startup control-flow decision.
-///
-/// Critically, `VersionCheckFailed` maps to `Halt`, NOT a fatal error: a transient
-/// distributor/network/parse failure during the startup check must keep `feral-setupd` alive as a
-/// BLE/polling retry surface. Propagating an error instead would exit the process and, under
-/// `Restart=always` + `StartLimitBurst=5`, crash-loop the daemon until systemd stops restarting it.
-/// The failure UI and durable-phase restoration already happened inside `check_and_update_system`.
-fn startup_update_check_outcome(result: UpdateCheckResult) -> StartupUpdateOutcome {
-    match result {
-        UpdateCheckResult::TooOldToUpgrade
-        | UpdateCheckResult::UpdateStarted
-        | UpdateCheckResult::UpdateInProgress
-        | UpdateCheckResult::VersionCheckFailed => StartupUpdateOutcome::Halt,
-        UpdateCheckResult::NoUpdateNeeded => StartupUpdateOutcome::Continue,
-    }
-}
-
-/// Follow-up after `check_and_update_system(UpdateMode::Available, …)` in the D-Bus path.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum AvailableSystemUpdateFollowUp {
-    /// No newer build: re-show the CURRENT canonical page so the TV leaves the transient
-    /// "Checking for updates..." URL the progress UI may have painted.
-    RestoreCurrentPage,
-    /// Blocking orchestration failed (CDP navigate, updater start, etc.).
-    LogFailure,
-    /// Core already drove the final TV screen (update started, reflash, classified fetch error).
-    NoOp,
-}
-
-/// Decide the follow-up for the D-Bus `system_update` path.
-///
-/// On `NoUpdateNeeded` the core did NOT navigate a final screen, so the TV is whatever was painted
-/// last — and because the version-check progress UI is transient (`navigate_transient_message`
-/// never records `page`), that may be the "Checking for updates..." URL even though the canonical
-/// page moved on. The caller therefore re-shows the CURRENT canonical page (read after the check):
-/// this both corrects a stuck transient screen and is inherently clobber-safe — it re-asserts the
-/// page another operation may have set during the check, never a stale pre-check snapshot. The
-/// progress task is already drained by `finish()` before this runs, so no later transient
-/// navigation can race the re-show. Other results already drove their own final screen → `NoOp`.
-fn available_system_update_follow_up(
-    result: &Result<UpdateCheckResult>,
-) -> AvailableSystemUpdateFollowUp {
-    match result {
-        Ok(UpdateCheckResult::NoUpdateNeeded) => AvailableSystemUpdateFollowUp::RestoreCurrentPage,
-        Err(_) => AvailableSystemUpdateFollowUp::LogFailure,
-        Ok(_) => AvailableSystemUpdateFollowUp::NoOp,
-    }
-}
-
-/// Canonical surface to re-show after a D-Bus `NoUpdateNeeded` check.
-///
-/// The version-check progress UI repaints Chromium with a transient "Checking for updates..."
-/// screen without recording `app_state.page`, so the no-update follow-up must actively re-show the
-/// real canonical surface or the TV stays stuck on that transient screen.
-#[derive(Debug, Clone, PartialEq, Eq)]
-enum RestorePageTarget {
-    Webapp,
-    Qrcode,
-    Message(String),
-    FactoryReset,
-    /// Nothing to restore. `ReflashingRequired` cannot be rebuilt from the stored page state (the
-    /// flashing-guide URL / QR content is not retained) and is unreachable on `NoUpdateNeeded`
-    /// anyway, since a too-old device returns `TooOldToUpgrade`. `None` has no canonical surface.
-    NoChange,
-}
-
-/// Decide which canonical page to re-show after a D-Bus no-update check.
-///
-/// `SystemUpgrade` is treated as a stale failure screen: once the manual retry cleared
-/// `update_failed`, re-painting the failure copy would be wrong, so it maps to the phase-appropriate
-/// surface (webapp when `Ready`, otherwise QR). Every other re-showable page maps back to itself so
-/// Chromium leaves the transient progress screen.
-fn restore_page_target(page: &Page, phase: SetupPhase) -> RestorePageTarget {
-    match page {
-        Page::SystemUpgrade(_) => match phase {
-            SetupPhase::Ready => RestorePageTarget::Webapp,
-            _ => RestorePageTarget::Qrcode,
-        },
-        Page::WebApp(_) => RestorePageTarget::Webapp,
-        Page::QRCode(_) => RestorePageTarget::Qrcode,
-        Page::Message(_, msg) => RestorePageTarget::Message(msg.clone()),
-        Page::FactoryReset(_) => RestorePageTarget::FactoryReset,
-        Page::ReflashingRequired(..) | Page::None(_) => RestorePageTarget::NoChange,
-    }
-}
-
-/// Shared system update logic that checks version status and optionally triggers an update.
-///
-/// This function handles:
-/// 1. Checking if the device is too old to auto-upgrade (shows reflashing QR code if so)
-/// 2. Checking if an update is needed based on the provided `UpdateMode`
-/// 3. Triggering the update process if needed
-///
-/// # Arguments
-/// * `app_state` - Application state
-/// * `chrome` - CDP connection for UI updates
-/// * `mode` - Controls whether to check for required updates only or any available update
-/// * `execution` - Controls whether operations block or run in background
-/// * `guard` - The already-acquired device-ownership guard. Callers MUST acquire this at their
-///   entry point (before any phase/UI/network side effects) so setup actions and OTA are serialized
-///   end-to-end, not just at this function's boundary. This closes the TOCTOU window where a load-only
-///   precheck could pass and a concurrent task then acquired ownership while the first path kept
-///   mutating Wi-Fi/phase/UI. The guard is released by RAII when this function returns, except on a
-///   non-blocking `UpdateStarted`, where ownership transfers into the background `update()` task and
-///   is held until the device reboots.
-///
-/// # Returns
-/// * `Ok(UpdateCheckResult)` indicating what action was taken
-/// * `Err` if a critical error occurred during the check (only possible with `Blocking` execution)
-async fn check_and_update_system(
-    app_state: &Arc<AppState>,
-    chrome: &Arc<Cdp>,
-    mode: UpdateMode,
-    execution: UpdateExecution,
-    guard: UpdateGuard,
-) -> Result<UpdateCheckResult> {
-    // Note: Failure latch clearing is handled by explicit retry call sites (BLE/D-Bus),
-    // not here. Startup flow needs to preserve persisted failure state so mobile app
-    // can observe update_failed across restarts (issue #447).
-
-    // Ownership was acquired by the caller before its side effects. Bind it here so RAII releases it
-    // on every return path (and so the non-blocking branch can move it into the update task).
-    let _guard = guard;
-
-    // Preserve the current phase to restore if no update is needed.
-    // Without this, durable phases (Ready, Pairing) would be lost during the
-    // mandatory no-update check, demoting paired devices to QR setup.
-    let phase_before_check = app_state.lifecycle.get();
-
-    // Store durable phase as PRE_FAILURE_PHASE before entering transient states.
-    // This ensures update_failed recovery can restore Ready/Pairing even though
-    // the failure occurs while in Updating (transient).
-    if phase_before_check.is_durable() && phase_before_check != SetupPhase::UpdateFailed {
-        use crate::persistent_state::PRE_FAILURE_PHASE;
-        app_state
-            .state_store
-            .set(PRE_FAILURE_PHASE, phase_before_check.as_str());
-        if let Err(e) = app_state.state_store.save() {
-            eprintln!("MAIN: Failed to persist pre-failure phase: {e:#?}");
-        }
-    }
-
-    app_state.lifecycle.set(SetupPhase::CheckingVersion);
-
-    let progress = VersionCheckProgress::start(execution, chrome.clone());
-
-    // BLE/NonBlocking flows hold the mobile notification open until this fetch resolves
-    // (`handle_connect_wifi` awaits the callback before replying), so cap them to a single
-    // attempt (worst case ≈ one request timeout) instead of the full ~34s retry budget that
-    // would violate the BLE response contract (docs/api-design.md "BLE response"). Blocking
-    // startup/D-Bus flows can wait, so they keep the full budget for resilience.
-    let retries = match execution {
-        UpdateExecution::Blocking => updater::RefreshRetries::Full,
-        UpdateExecution::NonBlocking => updater::RefreshRetries::Single,
-    };
-
-    // Force a fresh fetch first. This is a user-triggered/blocking check, so a failed live
-    // refresh must surface as classified copy rather than silently falling back to stale
-    // cached metadata (which could hide an outage or a newly raised minimum version).
-    if let Err(ve) = updater::refresh_remote_version(retries, progress.as_updater_progress()).await
-    {
-        progress.finish().await;
-        // Restore durable phase or reset to Idle on refresh failure (preserving paired state when
-        // distributor/network/parse errors occur during forced metadata refresh), without demoting
-        // a Ready that a pairing confirmation recorded while we were refreshing.
-        app_state.lifecycle.set(phase_after_inflight_check(
-            phase_before_check,
-            app_state.lifecycle.get(),
-        ));
-        return show_version_check_failure(app_state, chrome, execution, ve).await;
-    }
-    // First check if device is too old to auto-upgrade (reads the freshly refreshed cache)
-    match updater::is_too_old_to_upgrade(progress.as_updater_progress()).await {
-        Ok(true) => {
-            // Finish progress UI before reflashing screens; otherwise queued progress
-            // navigations can overwrite the QR/message we are about to show.
-            progress.finish().await;
-
-            // Device is too old, show reflashing QR code
-            if let Ok(Some(flashing_guide)) = updater::flashing_guide_url().await {
-                let current_version = app_state.current_version.clone();
-                let latest_version = updater::latest_version().await.unwrap_or_else(|e| {
-                    eprintln!("MAIN: latest_version for reflashing banner: {e:#?}");
-                    "Unknown".to_string()
-                });
-                let min_upgradeable = updater::min_upgradeable_version()
-                    .await
-                    .unwrap_or_else(|e| {
-                        eprintln!("MAIN: min_upgradeable_version for reflashing banner: {e:#?}");
-                        None
-                    })
-                    .unwrap_or_else(|| "Unknown".to_string());
-
-                match execution {
-                    UpdateExecution::Blocking => {
-                        show_reflashing_qrcode(
-                            app_state,
-                            chrome,
-                            &flashing_guide,
-                            &current_version,
-                            &latest_version,
-                            &min_upgradeable,
-                        )
-                        .await?;
-                    }
-                    UpdateExecution::NonBlocking => {
-                        task::spawn({
-                            let app_state = app_state.clone();
-                            let chrome = chrome.clone();
-                            async move {
-                                let _ = show_reflashing_qrcode(
-                                    &app_state,
-                                    &chrome,
-                                    &flashing_guide,
-                                    &current_version,
-                                    &latest_version,
-                                    &min_upgradeable,
-                                )
-                                .await;
-                            }
-                        });
-                    }
-                }
-            } else {
-                // Fallback to showing message without QR code if no flashing guide URL
-                match execution {
-                    UpdateExecution::Blocking => {
-                        show_message(chrome, app_state, constant::REFLASHING_REQUIRED_MSG).await?;
-                    }
-                    UpdateExecution::NonBlocking => {
-                        task::spawn({
-                            let app_state = app_state.clone();
-                            let chrome = chrome.clone();
-                            async move {
-                                let _ = show_message(
-                                    &chrome,
-                                    &app_state,
-                                    constant::REFLASHING_REQUIRED_MSG,
-                                )
-                                .await;
-                            }
-                        });
-                    }
-                }
-            }
-            app_state.lifecycle.set(SetupPhase::Idle);
-            return Ok(UpdateCheckResult::TooOldToUpgrade);
-        }
-        Ok(false) => {} // Device can be upgraded, continue
-        Err(e) => {
-            // Continue with the update check if this fails. This is safe because the forced
-            // `refresh_remote_version` above already succeeded (otherwise we'd have early-returned),
-            // so the cache is populated and this read is unlikely to fail; the subsequent
-            // `is_update_required`/`is_update_available` reads the same cache and surfaces any
-            // classified error there. The reflash gate is only skipped on a genuinely transient
-            // post-refresh failure, recovered on the next check.
-            eprintln!("MAIN: Error checking if device is too old: {e:#?}");
-        }
-    }
-
-    // Check if update is needed based on mode
-    let needs_update = match mode {
-        UpdateMode::Required => updater::is_update_required(progress.as_updater_progress()).await,
-        UpdateMode::Available => updater::is_update_available(progress.as_updater_progress()).await,
-    };
-
-    progress.finish().await;
-
-    match needs_update {
-        Ok(true) => {
-            match execution {
-                UpdateExecution::Blocking => {
-                    // For blocking execution during startup, keep setupd alive on permanent
-                    // failure so mobile polling can observe setup_phase=update_failed.
-                    // The guard remains held during update and auto-releases when this
-                    // branch completes.
-                    if let Err(e) = update(app_state.clone(), chrome.clone()).await {
-                        eprintln!("MAIN: Blocking update failed: {e:#?}");
-                        // update() already set setup_phase=update_failed and showed
-                        // error UI, then transitioned to QR. Return success so daemon
-                        // stays alive and mobile can poll the failure state.
-                    }
-                }
-                UpdateExecution::NonBlocking => {
-                    // For non-blocking execution (e.g., BLE-triggered updates), spawn
-                    // background task and transfer guard ownership into the task.
-                    // The guard remains held from acquisition to task completion,
-                    // preventing concurrent updates.
-                    let app_state_clone = app_state.clone();
-                    let chrome_clone = chrome.clone();
-                    task::spawn(async move {
-                        let _guard = _guard; // Transfer guard ownership into task
-                        let _ = update(app_state_clone, chrome_clone).await;
-                    });
-                }
-            }
-            Ok(UpdateCheckResult::UpdateStarted)
-        }
-        Ok(false) => {
-            if mode == UpdateMode::Available {
-                println!("MAIN: System update requested but no update available");
-            }
-            // Restore the durable phase captured before the transient CheckingVersion (Ready/Pairing
-            // preserved across the no-op check; non-durable collapses to Idle). phase_after_inflight_check
-            // additionally refuses to demote a Ready that a pairing confirmation recorded while this
-            // check was running, so the one-shot confirmation is not lost.
-            app_state.lifecycle.set(phase_after_inflight_check(
-                phase_before_check,
-                app_state.lifecycle.get(),
-            ));
-            Ok(UpdateCheckResult::NoUpdateNeeded)
-        }
-        Err(ve) => {
-            // Restore durable phase or reset to Idle on version-check failure (preserving paired
-            // state when distributor/network errors occur), without demoting a Ready that a pairing
-            // confirmation recorded while this check was running.
-            app_state.lifecycle.set(phase_after_inflight_check(
-                phase_before_check,
-                app_state.lifecycle.get(),
-            ));
-            show_version_check_failure(app_state, chrome, execution, ve).await
-        }
-    }
-}
-
-/// Render the classified version-check failure copy and report [`UpdateCheckResult::VersionCheckFailed`].
-///
-/// `Blocking` navigates inline (callers may await the screen); `NonBlocking` (BLE) spawns the
-/// navigation so the mobile response is not held on CDP.
-async fn show_version_check_failure(
-    app_state: &Arc<AppState>,
-    chrome: &Arc<Cdp>,
-    execution: UpdateExecution,
-    ve: updater::VersionFetchError,
-) -> Result<UpdateCheckResult> {
-    eprintln!("MAIN: Error checking for update: {ve:#?}");
-    let tv_msg = ve.kind().tv_message();
-    match execution {
-        UpdateExecution::Blocking => {
-            show_message(chrome, app_state, tv_msg).await?;
-        }
-        UpdateExecution::NonBlocking => {
-            let tv_msg = tv_msg.to_string();
-            task::spawn({
-                let app_state = app_state.clone();
-                let chrome = chrome.clone();
-                async move {
-                    let _ = show_message(&chrome, &app_state, &tv_msg).await;
-                }
-            });
-        }
-    }
-    Ok(UpdateCheckResult::VersionCheckFailed)
-}
-
-/// Drives TV “checking for updates” copy during distributor HTTP retries.
-///
-/// Only [`UpdateExecution::Blocking`] attaches a channel: BLE uses [`UpdateExecution::NonBlocking`]
-/// and passes `None` into the updater so CDP navigations are not on the mobile response path.
-/// When active, the sender is dropped and the task drained before final TV screens so progress
-/// cannot overwrite reflash, failure, or update UI.
-struct VersionCheckProgress {
-    tx: Option<updater::FetchProgressTx>,
-    task: Option<task::JoinHandle<()>>,
-}
-
-impl VersionCheckProgress {
-    #[must_use]
-    fn uses_tv_progress(execution: UpdateExecution) -> bool {
-        matches!(execution, UpdateExecution::Blocking)
-    }
-
-    // No `app_state`: progress navigations are transient and must NOT record the canonical
-    // page (see `navigate_transient_message`), so the receiver task only needs the CDP handle.
-    fn start(execution: UpdateExecution, chrome: Arc<Cdp>) -> Self {
-        if !Self::uses_tv_progress(execution) {
-            return Self {
-                tx: None,
-                task: None,
-            };
-        }
-
-        let (prog_tx, mut prog_rx) = mpsc::channel::<(u32, u32)>(16);
-        let progress_task = tokio::spawn(async move {
-            while let Some((_attempt, _max)) = prog_rx.recv().await {
-                if let Err(e) =
-                    navigate_transient_message(&chrome, constant::VERSION_CHECK_PROGRESS_TV_MSG)
-                        .await
-                {
-                    eprintln!("MAIN: version-check progress UI failed: {e:#?}");
-                }
-            }
-        });
-
-        Self {
-            tx: Some(prog_tx),
-            task: Some(progress_task),
-        }
-    }
-
-    fn as_updater_progress(&self) -> Option<updater::FetchProgressTx> {
-        self.tx.clone()
-    }
-
-    async fn finish(self) {
-        let Some(tx) = self.tx else {
-            return;
-        };
-        drop(tx);
-        if let Some(task) = self.task {
-            if let Err(join_err) = task.await {
-                eprintln!("MAIN: version-check progress task ended with: {join_err:#?}");
-            }
-        }
-    }
-}
-
-async fn show_webapp(app_state: &Arc<AppState>, chrome: &Arc<Cdp>) -> Result<()> {
-    let mut page = app_state.page.lock().await;
-
-    let webapp_url = constant::WEBAPP_URL;
-
-    // The static player is now a readiness-gated local service. We trust systemd
-    // for readiness and only keep the fast-path when Chromium is already on it.
-    let current_url = chrome.get_current_url().await?;
-    if current_url.starts_with(webapp_url) {
-        *page = Page::WebApp(unix_s());
-        return Ok(());
-    }
-
-    // This is to avoid Err Network Changed from Chrome
-    time::sleep(Duration::from_millis(constant::WIFI_WEBAPP_DELAY)).await;
-    chrome
-        .navigate(webapp_url)
-        .await
-        .with_context(|| format!("navigating to {webapp_url}"))?;
-
-    println!("MAIN: Navigated to {webapp_url}");
-    *page = Page::WebApp(unix_s());
-    Ok(())
-}
-
-async fn show_message(chrome: &Arc<Cdp>, app_state: &Arc<AppState>, message: &str) -> Result<()> {
-    let encoded = urlencoding::encode(message);
-    let message_url = format!("{}{}", constant::MSG_URL_PREFIX, encoded);
-    chrome
-        .navigate(&message_url)
-        .await
-        .with_context(|| format!("navigating to {message_url}"))?;
-    println!("MAIN: Navigated to {message_url}");
-
-    let mut page = app_state.page.lock().await;
-    *page = Page::Message(unix_s(), message.to_string());
-    Ok(())
-}
-
-/// Navigate the TV to a transient message WITHOUT recording it as the canonical `app_state.page`.
-///
-/// Used by the version-check progress UI. Because it never mutates `page`, a lagging progress
-/// navigation can no longer overwrite the page state set by a newer (possibly concurrent)
-/// transition — so the no-update restore guard reads a truthful page identity and cannot be
-/// tricked into clobbering a page another operation navigated to during the check.
-async fn navigate_transient_message(chrome: &Arc<Cdp>, message: &str) -> Result<()> {
-    let encoded = urlencoding::encode(message);
-    let message_url = format!("{}{}", constant::MSG_URL_PREFIX, encoded);
-    chrome
-        .navigate(&message_url)
-        .await
-        .with_context(|| format!("navigating to transient {message_url}"))?;
-    println!("MAIN: Navigated to transient {message_url}");
-    Ok(())
-}
-
-async fn show_system_upgrade(
-    chrome: &Arc<Cdp>,
-    app_state: &Arc<AppState>,
-    message: &str,
-) -> Result<()> {
-    let message_url = format!("{}{}", constant::MSG_URL_PREFIX, message);
-    chrome
-        .navigate(&message_url)
-        .await
-        .with_context(|| format!("navigating to {message_url}"))?;
-    println!("MAIN: Navigated to {message_url}");
-
-    let mut page = app_state.page.lock().await;
-    *page = Page::SystemUpgrade(unix_s());
-    Ok(())
-}
-
-async fn show_factory_reset(chrome: &Arc<Cdp>, app_state: &Arc<AppState>) -> Result<()> {
-    let message_url = format!(
-        "{}{}",
-        constant::MSG_URL_PREFIX,
-        constant::FACTORY_RESET_MSG
-    );
-    chrome
-        .navigate(&message_url)
-        .await
-        .with_context(|| format!("navigating to {message_url}"))?;
-    println!("MAIN: Navigated to {message_url}");
-
-    let mut page = app_state.page.lock().await;
-    *page = Page::FactoryReset(unix_s());
-    Ok(())
-}
-
-async fn wait_for_controld(timeout: Duration) -> Result<()> {
-    println!("MAIN: Waiting for controld connection...");
-
-    let wait_future = async {
-        loop {
-            match dbus_utils::check_dbus_connection(
-                constant::DBUS_CONTROLD_DESTINATION,
-                constant::DBUS_CONTROLD_OBJECT,
-            ) {
-                Ok(_) => break,
-                Err(e) => {
-                    println!("MAIN: controld not available yet: {e:#?}, retrying in 2 seconds...");
-                    time::sleep(Duration::from_secs(2)).await;
-                }
-            }
-        }
-    };
-
-    match time::timeout(timeout, wait_future).await {
-        Ok(_) => Ok(()),
-        Err(_) => Err(anyhow::anyhow!("Timeout waiting for controld connection")),
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
+    use app_state::{AppState, Page};
+    use connectivity::Connectivity;
+    use persistent_state::PersistentState;
+    use phase_logic::get_setup_phase;
+    use setup_lifecycle::{SetupLifecycle, SetupPhase};
+    use std::sync::Arc;
+    use std::sync::atomic::AtomicBool;
+    use tokio::sync::Mutex;
 
-    #[test]
-    fn classify_updater_message_marks_network_failures_transient() {
-        // Actual error from feral-updater.sh line 47
-        assert_eq!(
-            classify_updater_message("No network connection. Aborting update."),
-            UpdateErrorType::Transient
-        );
-        // Actual error from feral-system-update.sh line 147
-        assert_eq!(
-            classify_updater_message("Failed to retrieve content length for image."),
-            UpdateErrorType::Transient
-        );
-        // Download failures (fail-fast without retry in scripts)
-        // Used by both system and recovery update scripts
-        assert_eq!(
-            classify_updater_message("Failed to download OTA image."),
-            UpdateErrorType::Transient
-        );
-        assert_eq!(
-            classify_updater_message("Failed to download signature file."),
-            UpdateErrorType::Transient
-        );
-        assert_eq!(
-            classify_updater_message("Failed to download recovery ISO."),
-            UpdateErrorType::Transient
-        );
-    }
-
-    #[test]
-    fn classify_updater_message_marks_setupd_infrastructure_failures_transient() {
-        // Errors from updater.rs
-        assert_eq!(
-            classify_updater_message("Failed to start updater service: systemd error"),
-            UpdateErrorType::Transient
-        );
-        assert_eq!(
-            classify_updater_message("Failed to open /var/log/updaterd.log after retries"),
-            UpdateErrorType::Transient
-        );
-        assert_eq!(
-            classify_updater_message("updater closed channel without sending progress"),
-            UpdateErrorType::Transient
-        );
-    }
-
-    #[test]
-    fn classify_updater_message_marks_signing_and_image_failures_permanent() {
-        // Cryptographic failures (permanent - server issue)
-        assert_eq!(
-            classify_updater_message(
-                "Error: Signature verification failed for /var/tmp/ota/image.iso."
-            ),
-            UpdateErrorType::Permanent
-        );
-        // Image structure failures (permanent - corrupt image)
-        assert_eq!(
-            classify_updater_message("airootfs.sfs not found in image."),
-            UpdateErrorType::Permanent
-        );
-        // Filesystem operation failures (permanent - disk/permissions issue)
-        assert_eq!(
-            classify_updater_message(
-                "Failed to create snapshot '/.snapshots/@ota_prev'. Aborting."
-            ),
-            UpdateErrorType::Permanent
-        );
-        // Generic unknown errors
-        assert_eq!(
-            classify_updater_message("Unknown error occurred"),
-            UpdateErrorType::Permanent
-        );
-    }
-
-    #[test]
-    fn classify_updater_message_marks_lock_failures_transient() {
-        // Actual error from feral-updater.sh line 25
-        // Lock failures are transient because lock may be released by the time we retry
-        assert_eq!(
-            classify_updater_message(
-                "Exception: either Lock already held by another instance or some error happened."
-            ),
-            UpdateErrorType::Transient
-        );
-    }
-
-    #[test]
-    fn classify_updater_message_marks_unrecognized_as_permanent() {
-        assert_eq!(
-            classify_updater_message("something unexpected"),
-            UpdateErrorType::Permanent
-        );
-    }
-
-    #[test]
-    fn classify_updater_message_classifies_exception_err_by_command() {
-        // Network commands (curl, wget) are transient
-        assert_eq!(
-            classify_updater_message(
-                "EXCEPTION ERR: LINE=118 CMD=\"curl -u \"$auth_user:$auth_pass\" --silent --show-error -fL \"$ENDPOINT$IMAGE_URL\" -o \"$ZIP_FILE\"\""
-            ),
-            UpdateErrorType::Transient
-        );
-        assert_eq!(
-            classify_updater_message(
-                "EXCEPTION ERR: LINE=47 CMD=\"curl -su \"$auth_user:$auth_pass\" -f \"$API_URL\"\""
-            ),
-            UpdateErrorType::Transient
-        );
-        assert_eq!(
-            classify_updater_message(
-                "EXCEPTION ERR: LINE=200 CMD=\"wget --timeout=10 https://example.com/file.tar.gz\""
-            ),
-            UpdateErrorType::Transient
-        );
-
-        // Non-network commands are permanent
-        assert_eq!(
-            classify_updater_message(
-                "EXCEPTION ERR: LINE=99 CMD=\"mount -o loop /var/tmp/ota/image.iso /mnt/ota-iso\""
-            ),
-            UpdateErrorType::Permanent
-        );
-        assert_eq!(
-            classify_updater_message(
-                "EXCEPTION ERR: LINE=145 CMD=\"rsync -aAX --delete /mnt/ota-sfs/ /\""
-            ),
-            UpdateErrorType::Permanent
-        );
-        assert_eq!(
-            classify_updater_message("EXCEPTION ERR: LINE=208 CMD=\"mkinitcpio -P\""),
-            UpdateErrorType::Permanent
-        );
-    }
-
-    #[test]
-    fn set_and_get_setup_phase() {
+    /// Helper function to create test AppState with a given setup_phase
+    fn test_app_state(setup_phase: &str) -> Arc<AppState> {
         let lifecycle = SetupLifecycle::new();
+        lifecycle.set(match setup_phase {
+            "wifi_connecting" => SetupPhase::WifiConnecting,
+            "checking_version" => SetupPhase::CheckingVersion,
+            "updating" => SetupPhase::Updating,
+            "pairing" => SetupPhase::Pairing,
+            "ready" => SetupPhase::Ready,
+            "update_failed" => SetupPhase::UpdateFailed,
+            _ => SetupPhase::Idle,
+        });
+
         let temp_dir = tempfile::tempdir().unwrap();
         let state_file = temp_dir.path().join("state.txt");
         let state_store = PersistentState::new(state_file.to_str().unwrap()).unwrap();
 
-        let app_state = Arc::new(AppState {
+        Arc::new(AppState {
             device_id: "test".to_string(),
             branch: "test".to_string(),
             current_version: "1.0.0".to_string(),
@@ -2563,331 +180,10 @@ mod tests {
             auto_proceed: AtomicBool::new(false),
             lifecycle,
             update_in_progress: Arc::new(AtomicBool::new(false)),
-        });
-
-        assert_eq!(get_setup_phase(&app_state), "idle");
-
-        app_state.lifecycle.set(SetupPhase::WifiConnecting);
-        assert_eq!(get_setup_phase(&app_state), "wifi_connecting");
-
-        app_state.lifecycle.set(SetupPhase::CheckingVersion);
-        assert_eq!(get_setup_phase(&app_state), "checking_version");
-
-        app_state.lifecycle.set(SetupPhase::Updating);
-        assert_eq!(get_setup_phase(&app_state), "updating");
-
-        app_state.lifecycle.set(SetupPhase::UpdateFailed);
-        assert_eq!(get_setup_phase(&app_state), "update_failed");
-
-        app_state.lifecycle.set(SetupPhase::Pairing);
-        assert_eq!(get_setup_phase(&app_state), "pairing");
-
-        app_state.lifecycle.set(SetupPhase::Ready);
-        assert_eq!(get_setup_phase(&app_state), "ready");
+        })
     }
 
-    #[test]
-    fn build_device_info_includes_setup_phase() {
-        let lifecycle = SetupLifecycle::new();
-        lifecycle.set(SetupPhase::Updating);
-
-        let temp_dir = tempfile::tempdir().unwrap();
-        let state_file = temp_dir.path().join("state.txt");
-        let state_store = PersistentState::new(state_file.to_str().unwrap()).unwrap();
-
-        let app_state = Arc::new(AppState {
-            device_id: "test-device".to_string(),
-            branch: "main/stable".to_string(),
-            current_version: "1.2.3".to_string(),
-            state_store,
-            internet: tokio::runtime::Runtime::new()
-                .unwrap()
-                .block_on(async { Connectivity::spawn().await }),
-            page: Mutex::new(Page::None(0)),
-            auto_proceed: AtomicBool::new(false),
-            lifecycle,
-            update_in_progress: Arc::new(AtomicBool::new(false)),
-        });
-
-        let device_info = build_device_info(&app_state);
-        let parts: Vec<&str> = device_info.split('|').collect();
-
-        assert_eq!(parts.len(), 6);
-        assert_eq!(parts[0], "test-device");
-        assert_eq!(parts[1], ""); // topic_id (empty in test)
-        // parts[2] is internet status (true/false)
-        assert_eq!(parts[3], "main%2Fstable"); // branch with URL encoding
-        assert_eq!(parts[4], "1.2.3");
-        assert_eq!(parts[5], "updating");
-    }
-
-    #[test]
-    fn update_in_progress_guard_prevents_concurrent_updates() {
-        let temp_dir = tempfile::tempdir().unwrap();
-        let state_file = temp_dir.path().join("state.txt");
-        let state_store = PersistentState::new(state_file.to_str().unwrap()).unwrap();
-
-        let app_state = Arc::new(AppState {
-            device_id: "test".to_string(),
-            branch: "test".to_string(),
-            current_version: "1.0.0".to_string(),
-            state_store,
-            internet: tokio::runtime::Runtime::new()
-                .unwrap()
-                .block_on(async { Connectivity::spawn().await }),
-            page: Mutex::new(Page::None(0)),
-            auto_proceed: AtomicBool::new(false),
-            lifecycle: SetupLifecycle::new(),
-            update_in_progress: Arc::new(AtomicBool::new(false)),
-        });
-
-        // First acquisition should succeed
-        let guard1 = UpdateGuard::try_acquire(&app_state.update_in_progress);
-        assert!(guard1.is_some());
-
-        // Second acquisition should fail while first is held
-        let guard2 = UpdateGuard::try_acquire(&app_state.update_in_progress);
-        assert!(guard2.is_none());
-
-        // Drop first guard and reacquire should succeed
-        drop(guard1);
-        let guard3 = UpdateGuard::try_acquire(&app_state.update_in_progress);
-        assert!(guard3.is_some());
-    }
-
-    #[test]
-    fn update_guard_releases_flag_when_scope_ends() {
-        let flag = Arc::new(AtomicBool::new(false));
-
-        {
-            let _guard = UpdateGuard::try_acquire(&flag).expect("guard acquired");
-            assert!(flag.load(Ordering::Acquire));
-        }
-
-        assert!(!flag.load(Ordering::Acquire));
-        assert!(UpdateGuard::try_acquire(&flag).is_some());
-    }
-
-    /// Regression for PR #206 review 4483457995: setup actions and OTA must be serialized at the
-    /// operation level via the single ownership token, not a load-only precheck. This models the
-    /// race shape directly: once one path holds the guard (an OTA owns the device), every other
-    /// entry point (connect_wifi / keep_wifi / do_system_update / qr_switch / startup) fails to
-    /// acquire and must bail BEFORE any phase/UI/Wi-Fi side effects. Acquiring at entry (rather than
-    /// loading then acquiring later inside check_and_update_system) is what closes the TOCTOU window.
-    #[test]
-    fn device_ownership_is_exclusive_across_entry_points() {
-        let flag = Arc::new(AtomicBool::new(false));
-
-        // An OTA owns the device.
-        let ota_guard = UpdateGuard::try_acquire(&flag).expect("first acquire succeeds");
-
-        // Every other entry point's entry-time acquire is rejected, so they bail before side effects.
-        assert!(
-            UpdateGuard::try_acquire(&flag).is_none(),
-            "a second entry point must not acquire ownership while an OTA holds it",
-        );
-
-        // Once the OTA releases (completed/failed update), ownership is available again.
-        drop(ota_guard);
-        assert!(
-            UpdateGuard::try_acquire(&flag).is_some(),
-            "ownership must be reacquirable after the holder drops it",
-        );
-    }
-
-    /// The gate must track the real UpdateGuard lifecycle: blocked while a guard is held, allowed
-    /// again once it drops (so a finished/failed update re-opens setup actions).
-    #[test]
-    fn ble_action_gate_follows_update_guard_lifecycle() {
-        let flag = Arc::new(AtomicBool::new(false));
-        assert!(UpdateGuard::try_acquire(&flag).is_some());
-
-        {
-            let _guard = UpdateGuard::try_acquire(&flag).expect("guard acquired");
-            assert!(
-                UpdateGuard::try_acquire(&flag).is_none(),
-                "ownership is exclusive while a guard is held",
-            );
-        }
-
-        assert!(
-            UpdateGuard::try_acquire(&flag).is_some(),
-            "ownership is available again after the guard drops",
-        );
-    }
-
-    /// Regression for PR #206 review 4475889867: an offline reboot in UpdateFailed must short-circuit
-    /// to the recovery screen just like the online path, instead of falling through to QR while still
-    /// advertising setup_phase=update_failed. Only UpdateFailed triggers recovery; every other phase
-    /// proceeds with the normal startup flow, keeping both startup branches in lockstep.
-    #[test]
-    fn startup_recovery_only_for_update_failed() {
-        assert!(startup_requires_update_failed_recovery(
-            SetupPhase::UpdateFailed
-        ));
-        for phase in [
-            SetupPhase::Idle,
-            SetupPhase::WifiConnecting,
-            SetupPhase::CheckingVersion,
-            SetupPhase::Updating,
-            SetupPhase::Pairing,
-            SetupPhase::Ready,
-        ] {
-            assert!(
-                !startup_requires_update_failed_recovery(phase),
-                "phase {phase:?} must not trigger UpdateFailed recovery",
-            );
-        }
-    }
-
-    /// Regression for PR #206 review 4475817494: a BLE retry from update_failed restores a durable
-    /// phase, and a subsequent VersionCheckFailed must NOT demote it back to Idle. The wrapper
-    /// keeps durable phases (check_and_update_system already restored them) and only collapses
-    /// non-durable phases.
-    #[test]
-    fn ble_terminal_reset_preserves_durable_phase() {
-        assert_eq!(
-            ble_terminal_reset_phase(SetupPhase::Ready),
-            SetupPhase::Ready,
-        );
-        assert_eq!(
-            ble_terminal_reset_phase(SetupPhase::Pairing),
-            SetupPhase::Pairing,
-        );
-        assert_eq!(
-            ble_terminal_reset_phase(SetupPhase::UpdateFailed),
-            SetupPhase::UpdateFailed,
-        );
-    }
-
-    #[test]
-    fn ble_terminal_reset_collapses_non_durable_phase() {
-        assert_eq!(ble_terminal_reset_phase(SetupPhase::Idle), SetupPhase::Idle);
-        assert_eq!(
-            ble_terminal_reset_phase(SetupPhase::WifiConnecting),
-            SetupPhase::Idle,
-        );
-        assert_eq!(
-            ble_terminal_reset_phase(SetupPhase::CheckingVersion),
-            SetupPhase::Idle,
-        );
-        assert_eq!(
-            ble_terminal_reset_phase(SetupPhase::Updating),
-            SetupPhase::Idle,
-        );
-    }
-
-    /// Regression for PR #206 review 4482768497: the BLE no-update success path must not call the
-    /// blocking GetRelayerTopicID when a usable topic is already persisted (Pairing/Ready recovery);
-    /// it should return the persisted topic immediately. Only first-time setup (no/empty topic)
-    /// contacts controld.
-    #[test]
-    fn relayer_topic_fetch_only_when_no_persisted_topic() {
-        assert!(needs_relayer_topic_fetch(None));
-        assert!(needs_relayer_topic_fetch(Some("")));
-        assert!(!needs_relayer_topic_fetch(Some("topic-abc")));
-    }
-
-    /// Regression for PR #206 review 4482768497: a show_pairing_qr_code switch (either direction)
-    /// must be ignored while update_failed is latched, so the device stays on the failure surface
-    /// until an explicit retry clears it. Every other phase allows the switch.
-    #[test]
-    fn qr_switch_blocked_only_in_update_failed() {
-        assert!(qr_switch_blocked_by_update_failed(SetupPhase::UpdateFailed));
-        for phase in [
-            SetupPhase::Idle,
-            SetupPhase::WifiConnecting,
-            SetupPhase::CheckingVersion,
-            SetupPhase::Updating,
-            SetupPhase::Pairing,
-            SetupPhase::Ready,
-        ] {
-            assert!(
-                !qr_switch_blocked_by_update_failed(phase),
-                "phase {phase:?} must not block the QR switch",
-            );
-        }
-    }
-
-    /// Regression for PR #206 review 4483882810: a pairing confirmation must be recorded as a
-    /// durable Pairing -> Ready transition even when an OTA transiently masks Pairing, but must
-    /// never promote a first-run Idle (no topic), an already-Ready device, or a latched UpdateFailed.
-    #[test]
-    fn pairing_confirmation_promotes_only_paired_devices() {
-        // Steady-state Pairing always confirms (topic presence irrelevant; invariant guarantees it).
-        assert!(pairing_confirmation_promotes_to_ready(
-            SetupPhase::Pairing,
-            true
-        ));
-        assert!(pairing_confirmation_promotes_to_ready(
-            SetupPhase::Pairing,
-            false
-        ));
-
-        // Transient OTA phases masking a paired device (topic on disk) confirm; without a topic they
-        // are a first-run device mid-check and must not be promoted.
-        for masked in [
-            SetupPhase::WifiConnecting,
-            SetupPhase::CheckingVersion,
-            SetupPhase::Updating,
-        ] {
-            assert!(
-                pairing_confirmation_promotes_to_ready(masked, true),
-                "masked-paired phase {masked:?} should confirm"
-            );
-            assert!(
-                !pairing_confirmation_promotes_to_ready(masked, false),
-                "masked phase {masked:?} without a topic must not confirm"
-            );
-        }
-
-        // Never promote these, regardless of topic.
-        for phase in [
-            SetupPhase::Idle,
-            SetupPhase::Ready,
-            SetupPhase::UpdateFailed,
-        ] {
-            for has_topic in [true, false] {
-                assert!(
-                    !pairing_confirmation_promotes_to_ready(phase, has_topic),
-                    "phase {phase:?} (has_topic={has_topic}) must not confirm"
-                );
-            }
-        }
-    }
-
-    /// The in-flight check restore must preserve durable phases normally, but must NOT demote a
-    /// Ready that a pairing confirmation recorded while the check was running back to Pairing.
-    #[test]
-    fn inflight_check_restore_preserves_midcheck_ready() {
-        // Normal restores: captured durable phase wins; non-durable collapses to Idle.
-        assert_eq!(
-            phase_after_inflight_check(SetupPhase::Pairing, SetupPhase::CheckingVersion),
-            SetupPhase::Pairing
-        );
-        assert_eq!(
-            phase_after_inflight_check(SetupPhase::Ready, SetupPhase::CheckingVersion),
-            SetupPhase::Ready
-        );
-        assert_eq!(
-            phase_after_inflight_check(SetupPhase::Idle, SetupPhase::CheckingVersion),
-            SetupPhase::Idle
-        );
-
-        // A confirmation promoted the live phase to Ready mid-check: keep Ready, do not demote.
-        assert_eq!(
-            phase_after_inflight_check(SetupPhase::Pairing, SetupPhase::Ready),
-            SetupPhase::Ready
-        );
-
-        // A non-durable entry phase with a mid-check Ready is not demoted either (Ready is durable
-        // truth); but a non-durable entry without that promotion still collapses to Idle.
-        assert_eq!(
-            phase_after_inflight_check(SetupPhase::Idle, SetupPhase::Ready),
-            SetupPhase::Idle,
-            "only a captured Pairing is upgraded to a mid-check Ready"
-        );
-    }
+    // Integration tests - these test cross-module behavior
 
     #[test]
     fn qrcode_navigation_with_update_failed_phase_preserves_state() {
@@ -2915,35 +211,6 @@ mod tests {
 
         // Phase remains UpdateFailed
         assert_eq!(get_setup_phase(&app_state), "update_failed");
-    }
-
-    #[test]
-    fn build_device_info_exposes_update_failed_for_mobile_polling() {
-        let lifecycle = SetupLifecycle::new();
-        let temp_dir = tempfile::tempdir().unwrap();
-        let state_file = temp_dir.path().join("state.txt");
-        let store = PersistentState::new(state_file.to_str().unwrap()).unwrap();
-
-        lifecycle.set(SetupPhase::UpdateFailed);
-        lifecycle.persist(&store).unwrap();
-
-        let app_state = Arc::new(AppState {
-            device_id: "test-device".to_string(),
-            branch: "main/stable".to_string(),
-            current_version: "1.2.3".to_string(),
-            state_store: store,
-            internet: tokio::runtime::Runtime::new()
-                .unwrap()
-                .block_on(async { Connectivity::spawn().await }),
-            page: Mutex::new(Page::None(0)),
-            auto_proceed: AtomicBool::new(false),
-            lifecycle,
-            update_in_progress: Arc::new(AtomicBool::new(false)),
-        });
-
-        let device_info = build_device_info(&app_state);
-        let parts: Vec<&str> = device_info.split('|').collect();
-        assert_eq!(parts[5], "update_failed");
     }
 
     #[test]
@@ -3089,55 +356,6 @@ mod tests {
     }
 
     #[test]
-    fn update_attempt_error_classifies_from_raw_updater_message_before_context_wrap() {
-        let raw = "No network connection. Aborting update.";
-        let source = anyhow::anyhow!(raw);
-        let kind = classify_updater_message(&source.to_string());
-        let err = UpdateAttemptError {
-            source: source.context("update process failed"),
-            kind,
-        };
-
-        assert_eq!(err.kind, UpdateErrorType::Transient);
-        assert!(err.source.to_string().contains("update process failed"));
-        assert_eq!(
-            classify_updater_message(&err.source.to_string()),
-            UpdateErrorType::Permanent
-        );
-    }
-
-    fn test_app_state(setup_phase: &str) -> Arc<AppState> {
-        let lifecycle = SetupLifecycle::new();
-        lifecycle.set(match setup_phase {
-            "wifi_connecting" => SetupPhase::WifiConnecting,
-            "checking_version" => SetupPhase::CheckingVersion,
-            "updating" => SetupPhase::Updating,
-            "pairing" => SetupPhase::Pairing,
-            "ready" => SetupPhase::Ready,
-            "update_failed" => SetupPhase::UpdateFailed,
-            _ => SetupPhase::Idle,
-        });
-
-        let temp_dir = tempfile::tempdir().unwrap();
-        let state_file = temp_dir.path().join("state.txt");
-        let state_store = PersistentState::new(state_file.to_str().unwrap()).unwrap();
-
-        Arc::new(AppState {
-            device_id: "test".to_string(),
-            branch: "test".to_string(),
-            current_version: "1.0.0".to_string(),
-            state_store,
-            internet: tokio::runtime::Runtime::new()
-                .unwrap()
-                .block_on(async { Connectivity::spawn().await }),
-            page: Mutex::new(Page::None(0)),
-            auto_proceed: AtomicBool::new(false),
-            lifecycle,
-            update_in_progress: Arc::new(AtomicBool::new(false)),
-        })
-    }
-
-    #[test]
     fn ble_setup_success_transitions_pairing_to_ready() {
         let lifecycle = SetupLifecycle::new();
         let temp_dir = tempfile::tempdir().unwrap();
@@ -3205,222 +423,8 @@ mod tests {
         // Verify UpdateFailed is not restored (cleared by keep_wifi)
         assert_eq!(lifecycle_after_reboot.get(), SetupPhase::Idle);
     }
-}
 
-#[cfg(test)]
-mod version_check_progress_tests {
-    use super::{UpdateExecution, VersionCheckProgress};
-    use std::time::Duration;
-    use tokio::sync::mpsc;
-    use tokio::time::timeout;
-
-    #[test]
-    fn tv_progress_only_for_blocking_execution() {
-        assert!(VersionCheckProgress::uses_tv_progress(
-            UpdateExecution::Blocking
-        ));
-        assert!(!VersionCheckProgress::uses_tv_progress(
-            UpdateExecution::NonBlocking
-        ));
-    }
-
-    /// Mirrors `check_and_update_system`: ephemeral `Sender` clones must not outlive `finish`,
-    /// or the progress receiver task never sees the channel close.
-    #[tokio::test]
-    async fn progress_receiver_exits_after_all_senders_dropped() {
-        let (tx, mut rx) = mpsc::channel::<(u32, u32)>(16);
-        let per_updater_call = tx.clone();
-        drop(per_updater_call);
-
-        let receiver_task = tokio::spawn(async move { while rx.recv().await.is_some() {} });
-
-        drop(tx);
-        timeout(Duration::from_secs(1), receiver_task)
-            .await
-            .expect("receiver should finish within 1s")
-            .expect("receiver task join");
-    }
-}
-
-#[cfg(test)]
-mod available_system_update_follow_up_tests {
-    use super::{
-        AvailableSystemUpdateFollowUp, UpdateCheckResult, available_system_update_follow_up,
-    };
-    use anyhow::anyhow;
-
-    /// No update → re-show the current canonical page. The caller reads the CURRENT page (after the
-    /// check) and re-navigates to it, which both corrects a stuck transient progress screen and
-    /// re-asserts any page another operation set during the check (never clobbering it with a stale
-    /// snapshot). This is why the decision needs no prior/current comparison here.
-    #[test]
-    fn no_update_needed_restores_current_page() {
-        assert_eq!(
-            available_system_update_follow_up(&Ok(UpdateCheckResult::NoUpdateNeeded)),
-            AvailableSystemUpdateFollowUp::RestoreCurrentPage,
-        );
-    }
-
-    #[test]
-    fn update_started_is_no_op_for_dbus_follow_up() {
-        assert_eq!(
-            available_system_update_follow_up(&Ok(UpdateCheckResult::UpdateStarted)),
-            AvailableSystemUpdateFollowUp::NoOp,
-        );
-    }
-
-    #[test]
-    fn too_old_to_upgrade_is_no_op_for_dbus_follow_up() {
-        assert_eq!(
-            available_system_update_follow_up(&Ok(UpdateCheckResult::TooOldToUpgrade)),
-            AvailableSystemUpdateFollowUp::NoOp,
-        );
-    }
-
-    #[test]
-    fn version_check_failed_is_no_op_for_dbus_follow_up() {
-        assert_eq!(
-            available_system_update_follow_up(&Ok(UpdateCheckResult::VersionCheckFailed)),
-            AvailableSystemUpdateFollowUp::NoOp,
-        );
-    }
-
-    #[test]
-    fn orchestration_error_is_logged_only() {
-        assert_eq!(
-            available_system_update_follow_up(&Err(anyhow!("cdp navigate failed"))),
-            AvailableSystemUpdateFollowUp::LogFailure,
-        );
-    }
-}
-
-#[cfg(test)]
-mod startup_update_check_outcome_tests {
-    use super::{StartupUpdateOutcome, UpdateCheckResult, startup_update_check_outcome};
-
-    /// Regression for PR #206 review 4482497555: a failed mandatory startup version check must NOT
-    /// be fatal. on_startup_with_internet previously turned VersionCheckFailed into an Err that
-    /// propagated to main() and exited the process; with Restart=always + StartLimitBurst=5 a
-    /// transient distributor/network outage would crash-loop setupd until systemd gave up, killing
-    /// BLE advertising and the mobile retry surface. It must Halt (stay alive) instead.
-    #[test]
-    fn version_check_failed_halts_without_killing_daemon() {
-        assert_eq!(
-            startup_update_check_outcome(UpdateCheckResult::VersionCheckFailed),
-            StartupUpdateOutcome::Halt,
-        );
-    }
-
-    /// Too-old, update-started, and already-in-progress all left the daemon alive (early Ok) before
-    /// the refactor; lock that in so the helper keeps matching the historic control flow.
-    #[test]
-    fn updater_owned_results_halt_startup() {
-        for result in [
-            UpdateCheckResult::TooOldToUpgrade,
-            UpdateCheckResult::UpdateStarted,
-            UpdateCheckResult::UpdateInProgress,
-        ] {
-            assert_eq!(
-                startup_update_check_outcome(result),
-                StartupUpdateOutcome::Halt,
-                "result {result:?} must halt startup",
-            );
-        }
-    }
-
-    /// Only NoUpdateNeeded continues into topic_id fetch / surface selection.
-    #[test]
-    fn no_update_needed_continues_startup() {
-        assert_eq!(
-            startup_update_check_outcome(UpdateCheckResult::NoUpdateNeeded),
-            StartupUpdateOutcome::Continue,
-        );
-    }
-}
-
-#[cfg(test)]
-mod restore_page_target_tests {
-    use super::{Page, RestorePageTarget, SetupPhase, restore_page_target, unix_s};
-
-    /// Regression for PR #206 review 4475472053: after a manual D-Bus no-update check, the
-    /// transient "Checking for updates..." progress screen must be cleared by re-showing the real
-    /// canonical surface. The progress UI never records `app_state.page`, so the canonical page is
-    /// still the pre-check surface and must map back to itself (not a no-op, which left the TV stuck).
-    #[test]
-    fn webapp_canonical_page_reshows_webapp() {
-        assert_eq!(
-            restore_page_target(&Page::WebApp(unix_s()), SetupPhase::Ready),
-            RestorePageTarget::Webapp,
-        );
-    }
-
-    #[test]
-    fn qrcode_canonical_page_reshows_qrcode() {
-        assert_eq!(
-            restore_page_target(&Page::QRCode(unix_s()), SetupPhase::Pairing),
-            RestorePageTarget::Qrcode,
-        );
-    }
-
-    #[test]
-    fn message_canonical_page_reshows_same_message() {
-        let msg = "Setting things up".to_string();
-        assert_eq!(
-            restore_page_target(&Page::Message(unix_s(), msg.clone()), SetupPhase::Idle),
-            RestorePageTarget::Message(msg),
-        );
-    }
-
-    #[test]
-    fn factory_reset_canonical_page_reshows_factory_reset() {
-        assert_eq!(
-            restore_page_target(&Page::FactoryReset(unix_s()), SetupPhase::Idle),
-            RestorePageTarget::FactoryReset,
-        );
-    }
-
-    /// Stale failure screen: a prior UpdateFailed left `SystemUpgrade` as the canonical page, but
-    /// the retry that produced NoUpdateNeeded already cleared update_failed. Re-showing the failure
-    /// copy would be wrong, so route to the phase-appropriate surface instead.
-    #[test]
-    fn stale_system_upgrade_routes_by_phase() {
-        assert_eq!(
-            restore_page_target(&Page::SystemUpgrade(unix_s()), SetupPhase::Ready),
-            RestorePageTarget::Webapp,
-        );
-        assert_eq!(
-            restore_page_target(&Page::SystemUpgrade(unix_s()), SetupPhase::Pairing),
-            RestorePageTarget::Qrcode,
-        );
-        assert_eq!(
-            restore_page_target(&Page::SystemUpgrade(unix_s()), SetupPhase::Idle),
-            RestorePageTarget::Qrcode,
-        );
-    }
-
-    /// Reflashing cannot be rebuilt from stored page state and is unreachable on NoUpdateNeeded;
-    /// None has no surface. Both leave the page unchanged.
-    #[test]
-    fn reflashing_and_none_make_no_change() {
-        assert_eq!(
-            restore_page_target(
-                &Page::ReflashingRequired(unix_s(), "reflash".to_string()),
-                SetupPhase::Idle,
-            ),
-            RestorePageTarget::NoChange,
-        );
-        assert_eq!(
-            restore_page_target(&Page::None(unix_s()), SetupPhase::Idle),
-            RestorePageTarget::NoChange,
-        );
-    }
-}
-
-#[cfg(test)]
-mod phase_preservation_regression_tests {
-    use super::*;
-    use tempfile::NamedTempFile;
-
+    // Phase preservation regression tests for cross-module integration
     /// Regression test for PR #206 review finding: version-check failures should
     /// preserve durable phases (Ready, Pairing) instead of demoting to Idle.
     ///
@@ -3430,7 +434,7 @@ mod phase_preservation_regression_tests {
     /// Bug: Device gets demoted to Idle, shows QR instead of webapp.
     #[test]
     fn version_check_failure_preserves_ready_phase() {
-        let file = NamedTempFile::new().unwrap();
+        let file = tempfile::NamedTempFile::new().unwrap();
         let store = PersistentState::new(file.path().to_str().unwrap()).unwrap();
         let lifecycle = SetupLifecycle::new();
 
@@ -3473,7 +477,7 @@ mod phase_preservation_regression_tests {
     /// and restores it before the check so the snapshot is the durable Ready.
     #[test]
     fn connect_wifi_retry_no_update_keeps_ready_phase() {
-        let file = NamedTempFile::new().unwrap();
+        let file = tempfile::NamedTempFile::new().unwrap();
         let store = PersistentState::new(file.path().to_str().unwrap()).unwrap();
         let lifecycle = SetupLifecycle::new();
 
@@ -3520,7 +524,7 @@ mod phase_preservation_regression_tests {
     /// Same test for Pairing phase preservation on version-check failure.
     #[test]
     fn version_check_failure_preserves_pairing_phase() {
-        let file = NamedTempFile::new().unwrap();
+        let file = tempfile::NamedTempFile::new().unwrap();
         let store = PersistentState::new(file.path().to_str().unwrap()).unwrap();
         let lifecycle = SetupLifecycle::new();
 
@@ -3562,7 +566,7 @@ mod phase_preservation_regression_tests {
     /// where BLE reports "ready" but TV shows "update failed".
     #[test]
     fn dbus_retry_no_update_navigates_to_phase_driven_ui() {
-        let file = NamedTempFile::new().unwrap();
+        let file = tempfile::NamedTempFile::new().unwrap();
         let store = PersistentState::new(file.path().to_str().unwrap()).unwrap();
         let lifecycle = SetupLifecycle::new();
 
@@ -3635,78 +639,5 @@ mod phase_preservation_regression_tests {
                 "Transient phase {transient_phase:?} should reset to Idle on version check failure"
             );
         }
-    }
-
-    /// Regression test for PR #206 review finding (Jun 11): early refresh_remote_version
-    /// failure (lines 1638-1649) should preserve durable phases, not unconditionally set Idle.
-    ///
-    /// Scenario: Ready device, check_and_update_system runs, but refresh_remote_version
-    /// fails immediately (network/distributor/parse error) before we can check update status.
-    /// Expected: Device stays Ready, shows version-check failure UI.
-    /// Bug: Device demoted to Idle, paired device reports idle to mobile app.
-    #[test]
-    fn early_refresh_failure_preserves_ready_phase() {
-        let file = NamedTempFile::new().unwrap();
-        let store = PersistentState::new(file.path().to_str().unwrap()).unwrap();
-        let lifecycle = SetupLifecycle::new();
-
-        // Simulate a Ready device before update check
-        store.set(persistent_state::TOPIC_ID, "test-topic-ready");
-        lifecycle.set(SetupPhase::Ready);
-        lifecycle.persist(&store).unwrap();
-
-        assert_eq!(lifecycle.get(), SetupPhase::Ready);
-        let phase_before_check = lifecycle.get();
-
-        // Simulate entering CheckingVersion phase
-        lifecycle.set(SetupPhase::CheckingVersion);
-
-        // Early refresh_remote_version failure (network/distributor down, parse error)
-        // The fix (lines 1638-1649) should restore phase_before_check if durable
-        if phase_before_check.is_durable() {
-            lifecycle.set(phase_before_check);
-        } else {
-            lifecycle.set(SetupPhase::Idle);
-        }
-
-        // Verify: Ready is preserved, not demoted to Idle
-        assert_eq!(
-            lifecycle.get(),
-            SetupPhase::Ready,
-            "Early refresh failure should preserve Ready phase"
-        );
-    }
-
-    /// Same test for Pairing phase preservation on early refresh failure.
-    #[test]
-    fn early_refresh_failure_preserves_pairing_phase() {
-        let file = NamedTempFile::new().unwrap();
-        let store = PersistentState::new(file.path().to_str().unwrap()).unwrap();
-        let lifecycle = SetupLifecycle::new();
-
-        // Simulate a Pairing device (has topic_id, waiting for QR scan)
-        store.set(persistent_state::TOPIC_ID, "test-topic-pairing");
-        lifecycle.set(SetupPhase::Pairing);
-        lifecycle.persist(&store).unwrap();
-
-        assert_eq!(lifecycle.get(), SetupPhase::Pairing);
-        let phase_before_check = lifecycle.get();
-
-        // Simulate entering CheckingVersion phase
-        lifecycle.set(SetupPhase::CheckingVersion);
-
-        // Early refresh failure
-        if phase_before_check.is_durable() {
-            lifecycle.set(phase_before_check);
-        } else {
-            lifecycle.set(SetupPhase::Idle);
-        }
-
-        // Verify: Pairing is preserved
-        assert_eq!(
-            lifecycle.get(),
-            SetupPhase::Pairing,
-            "Early refresh failure should preserve Pairing phase"
-        );
     }
 }
